Migrate Magic Coat tag + other conflicts
