--- src/data/battler-tags.ts
@@ -1,3974 +1,0 @@
-import { globalScene } from "#app/global-scene";
-import { applyAbAttrs } from "./abilities/apply-ab-attrs";
-import { CommonBattleAnim } from "./animations/common-battle-anim";
-import { MoveChargeAnim } from "./animations/move-charge-anim";
-import { CommonAnim } from "#enums/common-anim";
-import { ChargeAnim } from "#enums/charge-anim";
-import { getMoveTargets, SelfStatusMove, type Move } from "#app/data/moves/move";
-import { applyMoveAttrs } from "#app/utils/move-utils";
-import { allMoves, allAbilities } from "#app/data/data-lists";
-import { StatusCategoryOnAllyAttr } from "./moves/move-attrs/status-category-on-ally-attr";
-import { HealOnAllyAttr } from "./moves/move-attrs/heal-on-ally-attr";
-import { MoveFlags } from "#enums/move-flags";
-import { MoveCategory } from "#enums/move-category";
-import { SpeciesFormChangeManualTrigger } from "./species-form-change-triggers/species-form-change-manual-trigger";
-import { getStatusEffectHealText } from "#app/data/status-effect";
-import { TerrainType } from "#enums/terrain-type";
-import { ElementalType } from "#enums/elemental-type";
-import type { Pokemon } from "#app/field/pokemon";
-import { MoveResult } from "#enums/move-result";
-import { HitResult } from "#enums/hit-result";
-import { getPokemonNameWithAffix } from "#app/messages";
-import { CommonAnimPhase } from "#app/phases/common-anim-phase";
-import { type MoveEffectPhase } from "#app/phases/move-effect-phase";
-import type { MovePhase } from "#app/phases/move-phase";
-import { ShowAbilityPhase } from "#app/phases/show-ability-phase";
-import { type StatStageChangeCallback } from "#app/phases/stat-stage-change-phase";
-import { StatStageChangePhase } from "#app/phases/stat-stage-change-phase";
-import i18next from "#app/plugins/i18n";
-import { BooleanHolder, getFrameMs, isNullOrUndefined, NumberHolder, toDmgValue } from "#app/utils";
-import { Abilities } from "#enums/abilities";
-import { BattlerTagType } from "#enums/battler-tag-type";
-import { MoveId } from "#enums/move-id";
-import { PokemonAnimType } from "#enums/pokemon-anim-type";
-import { Species } from "#enums/species";
-import { EFFECTIVE_STATS, getStatKey, Stat, type BattleStat, type EffectiveStat } from "#enums/stat";
-import { StatusEffect } from "#enums/status-effect";
-import { WeatherType } from "#enums/weather-type";
-import Overrides from "#app/overrides";
-import { BattlerTagLapseType } from "#enums/battler-tag-lapse-type";
-import { AbilityApplyMode } from "#enums/ability-apply-mode";
-import {
-  GulpMissileBattlerTagTypes,
-  MoveLockTagTypes,
-  RemoveTypeBattlerTagTypes,
-  SemiInvulnerableBattlerTagTypes,
-  TrappedBattlerTagTypes,
-} from "#app/utils/battler-tag-type-utils";
-import { AbAttrFlag } from "#enums/ab-attr-flag";
-import { PhaseId } from "#enums/phase-id";
-import { BattlerIndex } from "#enums/battler-index";
-import { MoveTarget } from "#enums/move-target";
-
-export class BattlerTag {
-  public tagType: BattlerTagType;
-  public lapseTypes: BattlerTagLapseType[];
-  public turnCount: number;
-  public sourceMoveId: MoveId;
-  public sourceId?: number;
-  public isBatonPassable: boolean;
-
-  constructor(
-    tagType: BattlerTagType,
-    lapseType: BattlerTagLapseType | BattlerTagLapseType[],
-    turnCount: number,
-    sourceMoveId?: MoveId,
-    sourceId?: number,
-    isBatonPassable: boolean = false,
-  ) {
-    this.tagType = tagType;
-    this.lapseTypes = Array.isArray(lapseType) ? lapseType : [lapseType];
-    this.turnCount = turnCount;
-    this.sourceMoveId = sourceMoveId!; // TODO: is this bang correct?
-    this.sourceId = sourceId;
-    this.isBatonPassable = isBatonPassable;
-  }
-
-  canAdd(_pokemon: Pokemon): boolean {
-    return true;
-  }
-
-  onAdd(_pokemon: Pokemon): void {}
-
-  onRemove(_pokemon: Pokemon): void {}
-
-  onOverlap(_pokemon: Pokemon): void {}
-
-  lapse(_pokemon: Pokemon, _lapseType: BattlerTagLapseType): boolean {
-    return --this.turnCount > 0;
-  }
-
-  apply(_pokemon: Pokemon, _simulated: boolean, ..._args: unknown[]): boolean {
-    return true;
-  }
-
-  getDescriptor(): string {
-    return "";
-  }
-
-  isSourceLinked(): boolean {
-    return false;
-  }
-
-  getMoveName(): string | null {
-    return this.sourceMoveId ? allMoves.get(this.sourceMoveId).name : null;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * This is meant to be inherited from by any battler tag with custom attributes
-   * @param source - The source {@linkcode BattlerTag}
-   */
-  loadTag(source: BattlerTag | any): void {
-    this.turnCount = source.turnCount;
-    this.sourceMoveId = source.sourceMoveId;
-    this.sourceId = source.sourceId;
-  }
-
-  /**
-   * Helper function that retrieves the source Pokemon object
-   * @returns The source {@linkcode Pokemon} or `null` if none is found
-   */
-  public getSourcePokemon(): Pokemon | null {
-    return this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-  }
-
-  isMysteryEncounterPostSummonTag(): this is MysteryEncounterPostSummonTag {
-    return false;
-  }
-
-  isTypeBoostTag(): this is TypeBoostTag {
-    return false;
-  }
-}
-
-export interface WeatherBattlerTag {
-  weatherTypes: WeatherType[];
-}
-
-export interface TerrainBattlerTag {
-  terrainTypes: TerrainType[];
-}
-
-export interface RestrictingBattlerTag {
-  getInterruptedText: (pokemon: Pokemon, moveId: MoveId) => string;
-  getSelectionDeniedText: (pokemon: Pokemon, moveId: MoveId) => string;
-}
-
-/**
- * Base class for tags that restrict the usage of moves. This effect is generally referred to as "disabling" a move
- * in-game. This is not to be confused with {@linkcode MoveId.DISABLE}.
- *
- * Descendants can override {@linkcode isMoveRestricted} to restrict moves that
- * match a condition. A restricted move gets cancelled before it is used. Players and enemies should not be allowed
- * to select restricted moves.
- */
-export abstract class MoveRestrictionBattlerTag extends BattlerTag implements RestrictingBattlerTag {
-  constructor(
-    tagType: BattlerTagType,
-    lapseType: BattlerTagLapseType | BattlerTagLapseType[],
-    turnCount: number,
-    sourceMoveId?: MoveId,
-    sourceId?: number,
-  ) {
-    super(tagType, lapseType, turnCount, sourceMoveId, sourceId);
-  }
-
-  /** @override */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.PRE_MOVE) {
-      // Cancel the affected pokemon's selected move
-      const phase = globalScene.getCurrentPhase() as MovePhase;
-      const move = phase.move;
-
-      if (this.isMoveRestricted(move.moveId, pokemon)) {
-        if (this.getInterruptedText(pokemon, move.moveId)) {
-          globalScene.queueMessage(this.getInterruptedText(pokemon, move.moveId));
-        }
-        phase.cancel();
-      }
-
-      return true;
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-
-  /**
-   * Gets whether this tag is restricting a move.
-   *
-   * @param moveId - {@linkcode MoveId} ID to check restriction for.
-   * @param user - The {@linkcode Pokemon} involved
-   * @returns `true` if the move is restricted by this tag, otherwise `false`.
-   */
-  public abstract isMoveRestricted(moveId: MoveId, user?: Pokemon): boolean;
-
-  /**
-   * Checks if this tag is restricting a move based on a user's decisions during the target selection phase
-   * @param _moveId - The {@linkcode MoveId | move} to check restriction for
-   * @param _user - The {@linkcode Pokemon | user} of the move
-   * @param _target - The {@linkcode Pokemon| target} of the move
-   * @returns `false` unless overridden by the child tag
-   */
-  public isMoveTargetRestricted(_moveId: MoveId, _user: Pokemon, _target: Pokemon): boolean {
-    return false;
-  }
-
-  /**
-   * Gets the text to display when the player attempts to select a move that is restricted by this tag.
-   *
-   * @param pokemon - {@linkcode Pokemon} for which the player is attempting to select the restricted move
-   * @param moveId - {@linkcode MoveId | move} that is having its selection denied
-   * @returns text to display when the player attempts to select the restricted move
-   */
-  abstract getSelectionDeniedText(pokemon: Pokemon, moveId: MoveId): string;
-
-  /**
-   * Gets the text to display when a move's execution is prevented as a result of the restriction.
-   * Because restriction effects also prevent selection of the move, this situation can only arise if a
-   * pokemon first selects a move, then gets outsped by a pokemon using a move that restricts the selected move.
-   *
-   * @param _pokemon - {@linkcode Pokemon} attempting to use the restricted move
-   * @param _moveId - The {@linkcode MoveId | move} being interrupted
-   * @returns text to display when the move is interrupted
-   */
-  abstract getInterruptedText(_pokemon: Pokemon, _moveId: MoveId): string;
-
-  /**
-   * Gets the last valid move from the pokemon's move history.
-   * @param pokemon {@linkcode Pokemon} to get the last valid move from
-   * @returns the last valid move from the pokemon's move history
-   */
-  public getLastValidMove(pokemon: Pokemon): Move | undefined {
-    const turnMove = pokemon
-      .getLastXMoves()
-      .find((m) => m.move.id !== MoveId.NONE && m.move.id !== MoveId.STRUGGLE && !m.virtual);
-
-    return turnMove?.move;
-  }
-}
-
-/**
- * Tag representing the "Throat Chop" effect. Pokemon with this tag cannot use sound-based moves.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Throat_Chop_(move) | Throat Chop}
- * @extends MoveRestrictionBattlerTag
- */
-export class ThroatChoppedTag extends MoveRestrictionBattlerTag {
-  constructor() {
-    super(
-      BattlerTagType.THROAT_CHOPPED,
-      [BattlerTagLapseType.TURN_END, BattlerTagLapseType.PRE_MOVE],
-      2,
-      MoveId.THROAT_CHOP,
-    );
-  }
-
-  /**
-   * Checks if a {@linkcode MoveId | move} is restricted by Throat Chop
-   * @param moveId - The {@linkcode MoveId | move} to check for sound-based restriction
-   * @returns `true` if the move is sound-based, `false` otherwise
-   */
-  override isMoveRestricted(moveId: MoveId): boolean {
-    return allMoves.get(moveId).hasFlag(MoveFlags.SOUND_MOVE);
-  }
-
-  /**
-   * Shows a message when the player attempts to select a move that is restricted by Throat Chop.
-   * @override
-   * @param _pokemon - The {@linkcode Pokemon} that is attempting to select the restricted move
-   * @param moveId - The {@linkcode MoveId | move} that is being restricted
-   * @returns the message to display when the player attempts to select the restricted move
-   */
-  override getSelectionDeniedText(_pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveCannotBeSelected", { moveName: allMoves.get(moveId).name });
-  }
-
-  /**
-   * Shows a message when a move is interrupted by Throat Chop.
-   * @override
-   * @param pokemon - The interrupted {@linkcode Pokemon}
-   * @param _moveId - The {@linkcode MoveId | move} that was interrupted
-   * @returns the message to display when the move is interrupted
-   */
-  override getInterruptedText(pokemon: Pokemon, _moveId: MoveId): string {
-    return i18next.t("battle:throatChopInterruptedMove", { pokemonName: getPokemonNameWithAffix(pokemon) });
-  }
-}
-
-/**
- * Tag representing the "disabling" effect performed by {@linkcode MoveId.DISABLE} and {@linkcode Abilities.CURSED_BODY}.
- * When the tag is added, the last-used move of the tag holder is set as the disabled move.
- */
-export class DisabledTag extends MoveRestrictionBattlerTag {
-  /** The move being disabled. Gets set when {@linkcode onAdd} is called for this tag. */
-  private moveId: MoveId = MoveId.NONE;
-
-  constructor(sourceId: number) {
-    super(
-      BattlerTagType.DISABLED,
-      [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END],
-      4,
-      MoveId.DISABLE,
-      sourceId,
-    );
-  }
-
-  /** @override */
-  override isMoveRestricted(moveId: MoveId): boolean {
-    return moveId === this.moveId;
-  }
-
-  /**
-   * @override
-   *
-   * Ensures that move history exists on `pokemon` and has a valid move. If so, sets the {@linkcode moveId} and shows a message.
-   * Otherwise the move ID will not get assigned and this tag will get removed next turn.
-   */
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    const lastValidMove = this.getLastValidMove(pokemon);
-    if (lastValidMove === undefined) {
-      return;
-    }
-
-    this.moveId = lastValidMove.id;
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:disabledOnAdd", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        moveName: allMoves.get(this.moveId).name,
-      }),
-    );
-  }
-
-  /** @override */
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:disabledLapse", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        moveName: allMoves.get(this.moveId).name,
-      }),
-    );
-  }
-
-  /** @override */
-  override getSelectionDeniedText(_pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabled", { moveName: allMoves.get(moveId).name });
-  }
-
-  /**
-   * @override
-   * @param pokemon - {@linkcode Pokemon} attempting to use the restricted move
-   * @param moveId - The {@linkcode MoveId | move} being interrupted
-   * @returns text to display when the move is interrupted
-   */
-  override getInterruptedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:disableInterruptedMove", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-    });
-  }
-
-  /** @override */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.moveId = source.moveId;
-  }
-}
-
-/**
- * Tag used by Gorilla Tactics to restrict the user to using only one move.
- * @extends MoveRestrictionBattlerTag
- */
-export class GorillaTacticsTag extends MoveRestrictionBattlerTag {
-  private moveId = MoveId.NONE;
-
-  constructor() {
-    super(BattlerTagType.GORILLA_TACTICS, BattlerTagLapseType.CUSTOM, 0);
-  }
-
-  /** @override */
-  override isMoveRestricted(moveId: MoveId): boolean {
-    return moveId !== this.moveId;
-  }
-
-  /**
-   * @override
-   * @param pokemon - The {@linkcode Pokemon} to check if the tag can be added
-   * @returns `true` if the pokemon has a valid move and no existing {@linkcode GorillaTacticsTag}; `false` otherwise
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    return this.getLastValidMove(pokemon) !== undefined && !pokemon.getTag(BattlerTagType.GORILLA_TACTICS);
-  }
-
-  /**
-   * Ensures that move history exists on {@linkcode Pokemon} and has a valid move.
-   * If so, sets the {@linkcode moveId} and increases the user's Attack by 50%.
-   * @override
-   * @param pokemon the {@linkcode Pokemon} to add the tag to
-   */
-  override onAdd(pokemon: Pokemon): void {
-    const lastValidMove = this.getLastValidMove(pokemon);
-
-    if (!lastValidMove) {
-      return;
-    }
-
-    this.moveId = lastValidMove.id;
-    pokemon.setStat(Stat.ATK, pokemon.getStat(Stat.ATK, false) * 1.5, false);
-  }
-
-  /**
-   * Loads the Gorilla Tactics Battler Tag along with its unique class variable moveId
-   * @override
-   * @param source Gorilla Tactics' {@linkcode BattlerTag} information
-   */
-  public override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.moveId = source.moveId;
-  }
-
-  /**
-   *
-   * @override
-   * @param pokemon n/a
-   * @param _moveId {@linkcode MoveId} ID of the move being denied
-   * @returns text to display when the move is denied
-   */
-  override getSelectionDeniedText(pokemon: Pokemon, _moveId: MoveId): string {
-    return i18next.t("battle:canOnlyUseMove", {
-      moveName: allMoves.get(this.moveId).name,
-      pokemonName: getPokemonNameWithAffix(pokemon),
-    });
-  }
-
-  override getInterruptedText(_pokemon: Pokemon, _moveId: MoveId): string {
-    return "";
-  }
-}
-
-/**
- * BattlerTag that represents the "recharge" effects of moves like Hyper Beam.
- */
-export class RechargingTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.RECHARGING, [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END], 2, sourceMoveId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    // Queue a placeholder move for the Pokemon to "use" next turn
-    pokemon.getMoveQueue().push({ move: SelfStatusMove.none(), targets: [], type: ElementalType.UNKNOWN });
-  }
-
-  /** Cancels the source's move this turn and queues a "__ must recharge!" message */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.PRE_MOVE) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:rechargingLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      (globalScene.getCurrentPhase() as MovePhase).cancel();
-      pokemon.getMoveQueue().shift();
-    }
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-/**
- * BattlerTag representing the rage effect where a Pokemon will gain +1 attack for each time it is hit
- * @extends BattlerTag
- */
-export class RageTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.RAGE, [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.AFTER_HIT], 1, MoveId.RAGE);
-  }
-
-  /**
-   * Grants +1atk to the owner of the tag while damaged while the tag is active and then immediately reapplies the tag
-   * @param pokemon the owner of the tag
-   * @param lapseType how this tag is lost
-   * @returns true if invoked with `AFTER_HIT` lapse type
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.AFTER_HIT) {
-      const lastAttackReceived = pokemon.turnData.attacksReceived[pokemon.turnData.attacksReceived.length - 1];
-      const damageReceived = lastAttackReceived?.damage ?? 0;
-      if (damageReceived > 0) {
-        globalScene.unshiftPhase(new StatStageChangePhase(pokemon.getBattlerIndex(), pokemon, [Stat.ATK], 1));
-      }
-      pokemon.addTag(BattlerTagType.RAGE, undefined, MoveId.RAGE, pokemon.id);
-      return true;
-    }
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-/**
- * BattlerTag representing the "charge phase" of Beak Blast.
- * Pokemon with this tag will inflict BURN status on any attacker that makes contact.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Beak_Blast_(move) | Beak Blast}
- */
-export class BeakBlastChargingTag extends BattlerTag {
-  constructor() {
-    super(
-      BattlerTagType.BEAK_BLAST_CHARGING,
-      [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END, BattlerTagLapseType.AFTER_HIT],
-      1,
-      MoveId.BEAK_BLAST,
-    );
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    // Play Beak Blast's charging animation
-    new MoveChargeAnim(ChargeAnim.BEAK_BLAST_CHARGING, this.sourceMoveId, pokemon).play();
-
-    // Queue Beak Blast's header message
-    globalScene.queueMessage(
-      i18next.t("moveTriggers:startedHeatingUpBeak", { pokemonName: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Inflicts `BURN` status on attackers that make contact, and causes this tag
-   * to be removed after the source makes a move (or the turn ends, whichever comes first)
-   * @param pokemon {@linkcode Pokemon} the owner of this tag
-   * @param lapseType {@linkcode BattlerTagLapseType} the type of functionality invoked in battle
-   * @returns `true` if invoked with the `AFTER_HIT` lapse type
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.AFTER_HIT) {
-      const phaseData = getMoveEffectPhaseData(pokemon);
-      if (phaseData?.move.hasFlag(MoveFlags.MAKES_CONTACT)) {
-        phaseData.attacker.trySetStatus(StatusEffect.BURN, true, pokemon);
-      }
-      return true;
-    }
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-/**
- * BattlerTag implementing Shell Trap's pre-move behavior.
- * Pokemon with this tag will act immediately after being hit by a physical move.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Shell_Trap_(move) | Shell Trap}
- */
-export class ShellTrapTag extends BattlerTag {
-  public activated: boolean = false;
-
-  constructor() {
-    super(BattlerTagType.SHELL_TRAP, [BattlerTagLapseType.TURN_END, BattlerTagLapseType.AFTER_HIT], 1);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    globalScene.queueMessage(
-      i18next.t("moveTriggers:setUpShellTrap", { pokemonName: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * "Activates" the shell trap, causing the tag owner to move next.
-   * @param pokemon {@linkcode Pokemon} the owner of this tag
-   * @param lapseType {@linkcode BattlerTagLapseType} the type of functionality invoked in battle
-   * @returns `true` if invoked with the `AFTER_HIT` lapse type
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.AFTER_HIT) {
-      const phaseData = getMoveEffectPhaseData(pokemon);
-
-      // Trap should only be triggered by opponent's Physical moves
-      if (phaseData?.move.category === MoveCategory.PHYSICAL && pokemon.isOpponent(phaseData.attacker)) {
-        const { turnManager } = globalScene.currentBattle;
-        this.activated = turnManager.preemptFightCommand((tc) => tc.pokemon === pokemon);
-      }
-
-      return true;
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-export class TrappedTag extends BattlerTag {
-  constructor(
-    tagType: BattlerTagType,
-    lapseType: BattlerTagLapseType,
-    turnCount: number,
-    sourceMoveId: MoveId,
-    sourceId: number,
-  ) {
-    super(tagType, lapseType, turnCount, sourceMoveId, sourceId, true);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    const source = globalScene.getPokemonById(this.sourceId!)!;
-    const move = allMoves.get(this.sourceMoveId);
-
-    const isGhost = pokemon.isOfType(ElementalType.GHOST);
-    const isTrapped = pokemon.getTag(...TrappedBattlerTagTypes);
-    const hasSubstitute = move.hitsSubstitute(source, pokemon);
-
-    return !isTrapped && !isGhost && (this.sourceMoveId === MoveId.G_MAX_TERROR || !hasSubstitute);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(this.getTrapMessage(pokemon));
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:trappedOnRemove", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        moveName: this.getMoveName(),
-      }),
-    );
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:trappedDesc");
-  }
-
-  override isSourceLinked(): boolean {
-    return true;
-  }
-
-  getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:trappedOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-}
-
-/**
- * BattlerTag implementing No Retreat's trapping effect.
- * This is treated separately from other trapping effects to prevent
- * Ghost-type Pokemon from being able to reuse the move.
- * @extends TrappedTag
- */
-class NoRetreatTag extends TrappedTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.NO_RETREAT, BattlerTagLapseType.CUSTOM, 0, MoveId.NO_RETREAT, sourceId);
-  }
-
-  /** overrides {@linkcode TrappedTag.apply}, removing the Ghost-type condition */
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.getTag(...TrappedBattlerTagTypes);
-  }
-}
-
-/**
- * BattlerTag that represents the {@link https://bulbapedia.bulbagarden.net/wiki/Flinch Flinch} status condition
- */
-export class FlinchedTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.FLINCHED, [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END], 0, sourceMoveId);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.isMax();
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    applyAbAttrs(AbAttrFlag.FLINCH_EFFECT, pokemon, false);
-  }
-
-  /**
-   * Cancels the Pokemon's next Move on the turn this tag is applied
-   * @param pokemon The {@linkcode Pokemon} with this tag
-   * @param lapseType The {@linkcode BattlerTagLapseType lapse type} used for this function call
-   * @returns `false` (This tag is always removed after applying its effects)
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.PRE_MOVE) {
-      (globalScene.getCurrentPhase() as MovePhase).cancel();
-      globalScene.queueMessage(
-        i18next.t("battlerTags:flinchedLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:flinchedDesc");
-  }
-}
-
-export class InterruptedTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.INTERRUPTED, BattlerTagLapseType.PRE_MOVE, 0, sourceMoveId);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !!pokemon.getTag(BattlerTagType.FLYING);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    pokemon.getMoveQueue().shift();
-    pokemon.pushMoveHistory({
-      move: SelfStatusMove.none(),
-      result: MoveResult.OTHER,
-      type: ElementalType.UNKNOWN,
-      targets: [],
-    });
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    (globalScene.getCurrentPhase() as MovePhase).cancel();
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-/**
- * BattlerTag that represents the {@link https://bulbapedia.bulbagarden.net/wiki/Confusion_(status_condition) Confusion} status condition
- */
-export class ConfusedTag extends BattlerTag {
-  public readonly ACTIVATION_CHANCE: number = 33;
-  constructor(turnCount: number, sourceMoveId: MoveId) {
-    super(BattlerTagType.CONFUSED, BattlerTagLapseType.MOVE, turnCount, sourceMoveId, undefined, true);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !globalScene.arena.hasTerrain(TerrainType.MISTY) || !pokemon.isGrounded();
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.unshiftPhase(new CommonAnimPhase(pokemon.getBattlerIndex(), undefined, CommonAnim.CONFUSION));
-    globalScene.queueMessage(
-      i18next.t("battlerTags:confusedOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:confusedOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    super.onOverlap(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:confusedOnOverlap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret =
-      (lapseType !== BattlerTagLapseType.CUSTOM && super.lapse(pokemon, lapseType))
-      || !isNullOrUndefined(Overrides.STATUS_ACTIVATION_OVERRIDE);
-
-    if (ret) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:confusedLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      globalScene.unshiftPhase(new CommonAnimPhase(pokemon.getBattlerIndex(), undefined, CommonAnim.CONFUSION));
-
-      const damage = this.getDamage(pokemon);
-      if (damage > 0) {
-        globalScene.queueMessage(i18next.t("battlerTags:confusedLapseHurtItself"));
-        pokemon.damageAndUpdate(damage);
-        pokemon.battleData.hitCount++;
-        (globalScene.getCurrentPhase() as MovePhase).cancel();
-      }
-    }
-    return ret;
-  }
-
-  /**
-   * Helper function for checking if Confusion activates and retrieving self-inflicted damage from confusion
-   * @param pokemon the confused Pokemon
-   * @returns the amount of damage inflicted
-   */
-  public getDamage(pokemon: Pokemon): number {
-    // 1/3 chance of hitting self with a 40 base power move
-    if (
-      (pokemon.randSeedInt(100) < this.ACTIVATION_CHANCE && Overrides.STATUS_ACTIVATION_OVERRIDE !== false)
-      || Overrides.STATUS_ACTIVATION_OVERRIDE === true
-    ) {
-      const atk = pokemon.getEffectiveStat(Stat.ATK, undefined, undefined, AbilityApplyMode.IGNORE);
-      const def = pokemon.getEffectiveStat(Stat.DEF, undefined, undefined, AbilityApplyMode.IGNORE);
-      return toDmgValue(
-        ((((2 * pokemon.level) / 5 + 2) * 40 * atk) / def / 50 + 2) * (pokemon.randSeedIntRange(85, 100) / 100),
-      );
-    }
-    return 0;
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:confusedDesc");
-  }
-}
-
-/**
- * Tag applied to the {@linkcode Move.DESTINY_BOND} user.
- * @extends BattlerTag
- * @see {@linkcode apply}
- */
-export class DestinyBondTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId, sourceId: number) {
-    super(BattlerTagType.DESTINY_BOND, BattlerTagLapseType.PRE_MOVE, 1, sourceMoveId, sourceId, true);
-  }
-
-  /**
-   * Lapses either before the user's move and does nothing
-   * or after receiving fatal damage. When the damage is fatal,
-   * the attacking Pokemon is taken down as well, unless it's a boss.
-   *
-   * @param pokemon Pokemon that is attacking the Destiny Bond user.
-   * @param lapseType CUSTOM or PRE_MOVE
-   * @returns `false` if the tag source fainted or one turn has passed since the application
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType !== BattlerTagLapseType.CUSTOM) {
-      return super.lapse(pokemon, lapseType);
-    }
-
-    const source = this.sourceId ? globalScene.getPokemonById(this.sourceId) : null;
-
-    if (!source || !source.isFainted() || !pokemon.isOnField() || source.getAlly() === pokemon) {
-      return false;
-    }
-
-    if (pokemon.isBossImmune()) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:destinyBondLapseIsBoss", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      return false;
-    }
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:destinyBondLapse", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(source),
-        pokemonNameWithAffix2: getPokemonNameWithAffix(pokemon),
-      }),
-    );
-    pokemon.damageAndUpdate(pokemon.hp, HitResult.ONE_HIT_KO, false, false, true);
-    return false;
-  }
-}
-
-export class InfatuatedTag extends BattlerTag {
-  public readonly ACTIVATION_CHANCE: number = 100 * (1 / 2);
-
-  constructor(sourceMoveId: number, sourceId: number) {
-    super(BattlerTagType.INFATUATED, BattlerTagLapseType.MOVE, 1, sourceMoveId, sourceId);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    if (this.sourceId) {
-      const pkm = globalScene.getPokemonById(this.sourceId);
-
-      if (pkm) {
-        return pokemon.isOppositeGender(pkm);
-      } else {
-        console.warn("canAdd: this.sourceId is not a valid pokemon id!", this.sourceId);
-        return false;
-      }
-    } else {
-      console.warn("canAdd: this.sourceId is undefined");
-      return false;
-    }
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:infatuatedOnAdd", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        sourcePokemonName: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-      }),
-    );
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    super.onOverlap(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:infatuatedOnOverlap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:infatuatedLapse", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          sourcePokemonName: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-        }),
-      );
-      globalScene.unshiftPhase(new CommonAnimPhase(pokemon.getBattlerIndex(), undefined, CommonAnim.ATTRACT));
-
-      if (
-        (pokemon.randSeedInt(100) < this.ACTIVATION_CHANCE && Overrides.STATUS_ACTIVATION_OVERRIDE !== false)
-        || Overrides.STATUS_ACTIVATION_OVERRIDE === true
-      ) {
-        globalScene.queueMessage(
-          i18next.t("battlerTags:infatuatedLapseImmobilize", {
-            pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          }),
-        );
-        (globalScene.getCurrentPhase() as MovePhase).cancel();
-      }
-    }
-
-    return ret;
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:infatuatedOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override isSourceLinked(): boolean {
-    return true;
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:infatuatedDesc");
-  }
-}
-
-export class SeedTag extends BattlerTag {
-  private sourceIndex: number;
-
-  constructor(sourceId: number) {
-    super(BattlerTagType.SEEDED, BattlerTagLapseType.TURN_END, 1, MoveId.LEECH_SEED, sourceId, true);
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.sourceIndex = source.sourceIndex;
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.isOfType(ElementalType.GRASS);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:seededOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-    this.sourceIndex = globalScene.getPokemonById(this.sourceId!)!.getBattlerIndex(); // TODO: are those bangs correct?
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      const source = pokemon.getOpponents().find((o) => o.getBattlerIndex() === this.sourceIndex);
-      if (source) {
-        const cancelled = new BooleanHolder(false);
-        applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-
-        if (!cancelled.value) {
-          globalScene.unshiftPhase(
-            new CommonAnimPhase(source.getBattlerIndex(), pokemon.getBattlerIndex(), CommonAnim.LEECH_SEED),
-          );
-
-          const damage = pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 8));
-          const reverseDrain = pokemon.hasAbilityWithAttr(AbAttrFlag.REVERSE_DRAIN, false);
-
-          globalScene.queuePokemonHeal(true, source.getBattlerIndex(), !reverseDrain ? damage : damage * -1, {
-            message: !reverseDrain
-              ? i18next.t("battlerTags:seededLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) })
-              : i18next.t("battlerTags:seededLapseShed", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-            showFullHpMessage: false,
-            skipAnim: true,
-          });
-        }
-      }
-    }
-
-    return ret;
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:seedDesc");
-  }
-}
-
-/**
- * BattlerTag representing the effects of {@link https://bulbapedia.bulbagarden.net/wiki/Powder_(move) | Powder}.
- * When the afflicted Pokemon uses a Fire-type move, the move is cancelled, and the
- * Pokemon takes damage equal to 1/4 of it's maximum HP (rounded down).
- */
-export class PowderTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.POWDER, [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END], 1);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    // "{Pokemon} is covered in powder!"
-    globalScene.queueMessage(
-      i18next.t("battlerTags:powderOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Applies Powder's effects before the tag owner uses a Fire-type move.
-   * Also causes the tag to expire at the end of turn.
-   * @param pokemon {@linkcode Pokemon} the owner of this tag
-   * @param lapseType {@linkcode BattlerTagLapseType} the type of lapse functionality to carry out
-   * @returns `true` if the tag should not expire after this lapse; `false` otherwise.
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.PRE_MOVE) {
-      const currPhase = globalScene.getCurrentPhase();
-      if (currPhase?.is<MovePhase>(PhaseId.MOVE)) {
-        const move = currPhase.move.getMove();
-        const weather = globalScene.arena.weather;
-        if (
-          pokemon.getMoveType(move) === ElementalType.FIRE
-          && !(weather && weather.weatherType === WeatherType.HEAVY_RAIN && !weather.isEffectSuppressed())
-        ) {
-          currPhase.fail();
-          currPhase.showMoveText();
-
-          globalScene.unshiftPhase(
-            new CommonAnimPhase(pokemon.getBattlerIndex(), pokemon.getBattlerIndex(), CommonAnim.POWDER),
-          );
-
-          const cancelDamage = new BooleanHolder(false);
-          applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelDamage);
-          if (!cancelDamage.value) {
-            pokemon.damageAndUpdate(Math.floor(pokemon.getMaxHp() / 4), HitResult.OTHER);
-          }
-
-          // "When the flame touched the powder\non the Pokémon, it exploded!"
-          globalScene.queueMessage(i18next.t("battlerTags:powderLapse", { moveName: move.name }));
-        }
-      }
-      return true;
-    } else {
-      return super.lapse(pokemon, lapseType);
-    }
-  }
-}
-
-export class NightmareTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.NIGHTMARE, BattlerTagLapseType.TURN_END, 1, MoveId.NIGHTMARE);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:nightmareOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    super.onOverlap(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:nightmareOnOverlap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:nightmareLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      globalScene.unshiftPhase(new CommonAnimPhase(pokemon.getBattlerIndex(), undefined, CommonAnim.CURSE)); // TODO: Update animation type
-
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-
-      if (!cancelled.value) {
-        pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 4));
-      }
-    }
-
-    return ret;
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:nightmareDesc");
-  }
-}
-
-/**
- * Locks the source into using a move consecutively for `turnCount - 1` turns. If the move fails or is interrupted
- * during any of these uses, this effect is removed.
- * @extends BattlerTag
- */
-export abstract class MoveLockTag extends BattlerTag {
-  protected lastTargets?: BattlerIndex[];
-
-  constructor(tagType: BattlerTagType, turnCount: number, sourceMoveId: MoveId) {
-    super(tagType, BattlerTagLapseType.AFTER_MOVE, turnCount, sourceMoveId);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    const moveQueue = pokemon.getMoveQueue();
-    moveQueue.splice(0, moveQueue.length);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.CUSTOM) {
-      return this.handleCustomLapse(pokemon);
-    } else {
-      return this.handleAfterMoveLapse(pokemon);
-    }
-  }
-
-  /**
-   * Queues the tag's source move on all turns except this tag's last turn
-   * if the tag holder's last move was successful. The queued move initially
-   * has no targets; the move targeting is instead resolved by this tag's custom lapse.
-   * @param pokemon the {@linkcode Pokemon} with this tag
-   * @returns `true` if the tag is to be retained after lapsing
-   */
-  protected handleAfterMoveLapse(pokemon: Pokemon): boolean {
-    if (!pokemon.getLastXMoves()?.[0]) {
-      return false;
-    }
-    const { move: lastMove, targets: lastTargets, result: lastMoveResult } = pokemon.getLastXMoves()[0];
-
-    // If this pokemon somehow used another move (e.g. via Dancer), don't advance this tag
-    if (![this.sourceMoveId, MoveId.NONE].includes(lastMove.id)) {
-      return true;
-    }
-
-    const ret =
-      super.lapse(pokemon, BattlerTagLapseType.AFTER_MOVE)
-      && lastTargets.length > 0
-      && lastMoveResult === MoveResult.SUCCESS;
-
-    if (ret) {
-      const move = allMoves.get(this.sourceMoveId);
-      this.lastTargets = lastTargets;
-      pokemon.getMoveQueue().push({ move, targets: [], ignorePP: true, type: pokemon.getMoveType(move) });
-    }
-
-    return ret;
-  }
-
-  /**
-   * Determines the next queued move's target during command selection
-   * @param pokemon the {@linkcode Pokemon} with this tag
-   */
-  protected handleCustomLapse(pokemon: Pokemon): boolean {
-    const queuedMove = pokemon.getMoveQueue()[0];
-
-    /**
-     * If the pokemon somehow doesn't have a queued move, abort
-     * target adjustment and remove this tag. This is just a failsafe;
-     * under normal circumstances, this should never happen.
-     */
-    if (!queuedMove) {
-      return false;
-    }
-
-    queuedMove.targets = this.getNextTargets(pokemon, queuedMove.move);
-    return true;
-  }
-
-  /**
-   * Obtains the target(s) for move actions queued by this tag
-   * @param pokemon the {@linkcode Pokemon} with this tag
-   * @param move the {@linkcode Move} queued by this tag
-   * @returns the target(s), by {@linkcode BattlerIndex}, for the next usage of this tag's move
-   */
-  protected getNextTargets(pokemon: Pokemon, move: Move): BattlerIndex[] {
-    if (move.moveTarget === MoveTarget.RANDOM_NEAR_ENEMY) {
-      return getMoveTargets(pokemon, move.id).targets;
-    } else {
-      // Failsafe if `this.lastTargets` has somehow not been set
-      if (!this.lastTargets?.length) {
-        this.lastTargets = pokemon.isPlayer() ? [BattlerIndex.ENEMY] : [BattlerIndex.PLAYER];
-      }
-
-      // Note: this assumes the locked move is single-target
-      const lastTarget = globalScene.getFieldPokemonByBattlerIndex(this.lastTargets[0]);
-      const adjacentIndex = this.lastTargets[0] + (this.lastTargets[0] % 2 === 0 ? 1 : -1);
-      const adjacentTarget = globalScene.getFieldPokemonByBattlerIndex(adjacentIndex);
-
-      if (
-        !lastTarget?.isActive(true)
-        && globalScene.currentBattle.double
-        && adjacentTarget?.isActive(true)
-        && adjacentTarget !== pokemon
-      ) {
-        return [adjacentIndex];
-      } else {
-        return this.lastTargets;
-      }
-    }
-  }
-
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.lastTargets = source.lastTargets;
-  }
-}
-
-/**
- * Puts the source {@linkcode Pokemon} into a "frenzy", locking them into using the tag's
- * move for `turnCount - 1` turns. If this effect isn't interrupted,
- * the source {@linkcode Pokemon} becomes confused at the end of those turns.
- * @extends MoveLockTag
- * @see {@linkcode FrenzyAttr}
- */
-export class FrenzyTag extends MoveLockTag {
-  constructor(turnCount: number, sourceMoveId: MoveId) {
-    super(BattlerTagType.FRENZY, turnCount, sourceMoveId);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    if (this.turnCount <= 0) {
-      // Only add CONFUSED tag if a disruption occurs on the final confusion-inducing turn of FRENZY
-      pokemon.addTag(BattlerTagType.CONFUSED, pokemon.randSeedIntRange(2, 4));
-    }
-  }
-}
-
-/**
- * Applies the move-locking effect of {@link https://bulbapedia.bulbagarden.net/wiki/Rollout_(move) Rollout}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Ice_Ball_(move) Ice Ball}.
- * Also defines a power multiplier for the respective move based on
- * the tag's turn count.
- * @extends MoveLockTag
- */
-export class RollingTag extends MoveLockTag {
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.ROLLING, 5, sourceMoveId);
-  }
-
-  public get powerMultiplier() {
-    return Math.pow(2, 5 - this.turnCount);
-  }
-}
-
-/**
- * Puts the source {@linkcode Pokemon} into an uproar, locking them into using
- * Uproar for 2 turns after the initial usage and preventing all
- * Pokemon on the field from sleeping. All Pokemon on the field also
- * wake up when this tag is added.
- * @extends MoveLockTag
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Uproar_(move) Uproar}
- */
-export class UproarTag extends MoveLockTag {
-  constructor() {
-    super(BattlerTagType.UPROAR, 3, MoveId.UPROAR);
-  }
-
-  /**
-   * Plays a "started an uproar" message, then wakes up all active Pokemon
-   * @param pokemon the {@linkcode Pokemon} with this tag
-   */
-  override onAdd(pokemon: Pokemon): void {
-    // "{pokemonNameWithAffix} caused an uproar!"
-    globalScene.queueMessage(
-      i18next.t("battlerTags:uproarOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-
-    // Wake up all sleeping Pokemon on the field
-    globalScene.getField(true).forEach((p) => {
-      if (p.hasStatusEffect(StatusEffect.SLEEP, false, true)) {
-        p.resetStatus();
-        // "The uproar woke {pokemonNameWithAffix}!"
-        globalScene.queueMessage(
-          i18next.t("battlerTags:uproarOnCureSleep", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        );
-      }
-    });
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    // "{pokemonNameWithAffix} calmed down."
-    globalScene.queueMessage(
-      i18next.t("battlerTags:uproarOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-
-    super.onRemove(pokemon);
-  }
-
-  /**
-   * Prevents Pokemon on the field from falling asleep
-   * @param pokemon the {@linkcode Pokemon} with this tag
-   * @param simulated if `true`, suppresses changes to game state
-   * @param affectedPokemon the {@linkcode Pokemon} to be afflicted with sleep
-   * @param preventSleep a {@linkcode BooleanHolder} which, if `true`, cancels attempts to afflict sleep
-   * @returns `true`
-   */
-  override apply(
-    _pokemon: Pokemon,
-    simulated: boolean,
-    affectedPokemon: Pokemon,
-    preventSleep: BooleanHolder,
-  ): boolean {
-    if (!simulated) {
-      // "But the uproar kept {pokemonNameWithAffix} awake!"
-      globalScene.queueMessage(
-        i18next.t("battlerTags:uproarOnPreventSleep", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(affectedPokemon),
-        }),
-      );
-    }
-
-    preventSleep.value = true;
-    return true;
-  }
-}
-
-/**
- * Applies the effects of the move Encore onto the target Pokemon
- * Encore forces the target Pokemon to use its most-recent move for 3 turns
- */
-export class EncoreTag extends MoveRestrictionBattlerTag {
-  public moveId: MoveId;
-
-  constructor(sourceId: number) {
-    super(
-      BattlerTagType.ENCORE,
-      [BattlerTagLapseType.CUSTOM, BattlerTagLapseType.AFTER_MOVE],
-      3,
-      MoveId.ENCORE,
-      sourceId,
-    );
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.moveId = source.moveId as MoveId;
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    const lastMove = pokemon.getLastXMoves(-1).filter((mv) => !mv.virtual)[0];
-    this.moveId = lastMove?.move.id;
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:encoreOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-
-    const { turnManager } = globalScene.currentBattle;
-
-    const movesetMove = pokemon.getMoveset().find((m) => m.moveId === this.moveId);
-    if (movesetMove) {
-      turnManager.tryReplaceMove(pokemon, movesetMove, lastMove?.targets ?? []);
-    }
-  }
-
-  /**
-   * If the encored move has run out of PP, Encore ends early. Otherwise, Encore lapses based on the AFTER_MOVE battler tag lapse type.
-   * @returns `true` to persist | `false` to end and be removed
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.CUSTOM) {
-      const encoredMove = pokemon.getMoveset().find((m) => m.moveId === this.moveId);
-      if (encoredMove && encoredMove?.getPpRatio() > 0) {
-        return true;
-      }
-      return false;
-    } else {
-      return super.lapse(pokemon, lapseType);
-    }
-  }
-
-  /**
-   * Checks if the move matches the moveId stored within the tag and returns a boolean value
-   * @param moveId {@linkcode MoveId} the move selected
-   * @param user N/A
-   * @returns `true` if the move does not match with the moveId stored and as a result, restricted
-   */
-  override isMoveRestricted(moveId: MoveId, _user?: Pokemon): boolean {
-    if (moveId !== this.moveId) {
-      return true;
-    }
-    return false;
-  }
-
-  override getSelectionDeniedText(_pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabled", { moveName: allMoves.get(moveId).name });
-  }
-
-  override getInterruptedText(_pokemon: Pokemon, _moveId: MoveId): string {
-    return "";
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:encoreOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-export class HelpingHandTag extends BattlerTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.HELPING_HAND, BattlerTagLapseType.TURN_END, 1, MoveId.HELPING_HAND, sourceId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    globalScene.queueMessage(
-      i18next.t("battlerTags:helpingHandOnAdd", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-        pokemonName: getPokemonNameWithAffix(pokemon),
-      }),
-    );
-  }
-}
-
-/**
- * Applies the Ingrain tag to a pokemon
- * @extends TrappedTag
- */
-export class IngrainTag extends TrappedTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.INGRAIN, BattlerTagLapseType.TURN_END, 1, MoveId.INGRAIN, sourceId);
-  }
-
-  /**
-   * Check if the Ingrain tag can be added to the pokemon
-   * @param pokemon {@linkcode Pokemon} The pokemon to check if the tag can be added to
-   * @returns boolean True if the tag can be added, false otherwise
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    const isTrapped = pokemon.getTag(BattlerTagType.TRAPPED);
-
-    return !isTrapped;
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queuePokemonHeal(true, pokemon.getBattlerIndex(), toDmgValue(pokemon.getMaxHp() / 16), {
-        message: i18next.t("battlerTags:ingrainLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      });
-    }
-
-    return ret;
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:ingrainOnTrap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:ingrainDesc");
-  }
-}
-
-/**
- * Octolock traps the target pokemon and reduces its DEF and SPDEF by one stage at the
- * end of each turn.
- */
-export class OctolockTag extends TrappedTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.OCTOLOCK, BattlerTagLapseType.TURN_END, 1, MoveId.OCTOLOCK, sourceId);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const shouldLapse = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (shouldLapse) {
-      globalScene.unshiftPhase(
-        new StatStageChangePhase(pokemon.getBattlerIndex(), null, [Stat.DEF, Stat.SPDEF], -1, { bypassReflect: true }),
-      );
-      return true;
-    }
-
-    return false;
-  }
-}
-
-export class AquaRingTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.AQUA_RING, BattlerTagLapseType.TURN_END, 1, MoveId.AQUA_RING, undefined, true);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:aquaRingOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queuePokemonHeal(true, pokemon.getBattlerIndex(), toDmgValue(pokemon.getMaxHp() / 16), {
-        message: i18next.t("battlerTags:aquaRingLapse", {
-          moveName: this.getMoveName(),
-          pokemonName: getPokemonNameWithAffix(pokemon),
-        }),
-      });
-    }
-
-    return ret;
-  }
-}
-
-/** Tag used to allow moves that interact with {@link MoveId.MINIMIZE} to function */
-export class MinimizeTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.MINIMIZED, BattlerTagLapseType.TURN_END, 1, MoveId.MINIMIZE);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.isMax();
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    //If a pokemon dynamaxes they lose minimized status
-    if (pokemon.isMax()) {
-      return false;
-    }
-    return lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-  }
-}
-
-export class DrowsyTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.DROWSY, BattlerTagLapseType.TURN_END, 2, MoveId.YAWN);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !globalScene.arena.hasTerrain(TerrainType.ELECTRIC) || !pokemon.isGrounded();
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:drowsyOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (!super.lapse(pokemon, lapseType)) {
-      pokemon.trySetStatus(StatusEffect.SLEEP, true);
-      return false;
-    }
-
-    return true;
-  }
-
-  override getDescriptor(): string {
-    return i18next.t("battlerTags:drowsyDesc");
-  }
-}
-
-export abstract class DamagingTrapTag extends TrappedTag {
-  private commonAnim: CommonAnim;
-
-  constructor(
-    tagType: BattlerTagType,
-    commonAnim: CommonAnim,
-    turnCount: number,
-    sourceMoveId: MoveId,
-    sourceId: number,
-  ) {
-    super(tagType, BattlerTagLapseType.TURN_END, turnCount, sourceMoveId, sourceId);
-
-    this.commonAnim = commonAnim;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.commonAnim = source.commonAnim as CommonAnim;
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.getTag(...TrappedBattlerTagTypes) && !pokemon.getTag(BattlerTagType.SUBSTITUTE);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:damagingTrapLapse", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          moveName: this.getMoveName(),
-        }),
-      );
-      globalScene.unshiftPhase(new CommonAnimPhase(pokemon.getBattlerIndex(), undefined, this.commonAnim));
-
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-
-      if (!cancelled.value) {
-        pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 8));
-      }
-    }
-
-    return ret;
-  }
-}
-
-export class BindTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.BIND, CommonAnim.BIND, turnCount, MoveId.BIND, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:bindOnTrap", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      sourcePokemonName: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-      moveName: this.getMoveName(),
-    });
-  }
-}
-
-export class WrapTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.WRAP, CommonAnim.WRAP, turnCount, MoveId.WRAP, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:wrapOnTrap", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      sourcePokemonName: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-    });
-  }
-}
-
-export abstract class VortexTrapTag extends DamagingTrapTag {
-  constructor(
-    tagType: BattlerTagType,
-    commonAnim: CommonAnim,
-    turnCount: number,
-    sourceMoveId: MoveId,
-    sourceId: number,
-  ) {
-    super(tagType, commonAnim, turnCount, sourceMoveId, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:vortexOnTrap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-}
-
-export class FireSpinTag extends VortexTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.FIRE_SPIN, CommonAnim.FIRE_SPIN, turnCount, MoveId.FIRE_SPIN, sourceId);
-  }
-}
-
-/**
- * Used for G-Max Centiferno that leaves a fire spin
- * that persists even on the user leaving the field
- */
-export class GMaxFireSpinTag extends FireSpinTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(turnCount, sourceId);
-    this.tagType = BattlerTagType.G_MAX_FIRE_SPIN;
-    this.sourceMoveId = MoveId.G_MAX_CENTIFERNO;
-  }
-
-  override isSourceLinked(): boolean {
-    return false;
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.getTag(...TrappedBattlerTagTypes);
-  }
-}
-
-export class WhirlpoolTag extends VortexTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.WHIRLPOOL, CommonAnim.WHIRLPOOL, turnCount, MoveId.WHIRLPOOL, sourceId);
-  }
-}
-
-export class ClampTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.CLAMP, CommonAnim.CLAMP, turnCount, MoveId.CLAMP, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:clampOnTrap", {
-      sourcePokemonNameWithAffix: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-      pokemonName: getPokemonNameWithAffix(pokemon),
-    });
-  }
-}
-
-export class SandTombTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.SAND_TOMB, CommonAnim.SAND_TOMB, turnCount, MoveId.SAND_TOMB, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:sandTombOnTrap", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: this.getMoveName(),
-    });
-  }
-}
-
-/**
- * Used for G-Max Sandblast that leaves a sand tomb
- * that persists even on the user leaving the field
- */
-export class GMaxSandTombTag extends SandTombTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(turnCount, sourceId);
-    this.tagType = BattlerTagType.G_MAX_SAND_TOMB;
-    this.sourceMoveId = MoveId.G_MAX_SANDBLAST;
-  }
-
-  override isSourceLinked(): boolean {
-    return false;
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.getTag(...TrappedBattlerTagTypes);
-  }
-}
-
-export class MagmaStormTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.MAGMA_STORM, CommonAnim.MAGMA_STORM, turnCount, MoveId.MAGMA_STORM, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:magmaStormOnTrap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-}
-
-export class SnapTrapTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.SNAP_TRAP, CommonAnim.SNAP_TRAP, turnCount, MoveId.SNAP_TRAP, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:snapTrapOnTrap", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-}
-
-export class ThunderCageTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.THUNDER_CAGE, CommonAnim.THUNDER_CAGE, turnCount, MoveId.THUNDER_CAGE, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:thunderCageOnTrap", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      sourcePokemonNameWithAffix: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-    });
-  }
-}
-
-export class InfestationTag extends DamagingTrapTag {
-  constructor(turnCount: number, sourceId: number) {
-    super(BattlerTagType.INFESTATION, CommonAnim.INFESTATION, turnCount, MoveId.INFESTATION, sourceId);
-  }
-
-  override getTrapMessage(pokemon: Pokemon): string {
-    return i18next.t("battlerTags:infestationOnTrap", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      sourcePokemonNameWithAffix: getPokemonNameWithAffix(globalScene.getPokemonById(this.sourceId!) ?? undefined), // TODO: is that bang correct?
-    });
-  }
-}
-
-export class ProtectedTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId, tagType: BattlerTagType = BattlerTagType.PROTECTED) {
-    super(tagType, BattlerTagLapseType.TURN_END, 0, sourceMoveId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:protectedOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (move.checkFlag(MoveFlags.IGNORE_PROTECT, attacker, pokemon)) {
-      return false;
-    }
-
-    if (!simulated) {
-      new CommonBattleAnim(CommonAnim.PROTECT, pokemon).play();
-      globalScene.queueMessage(
-        i18next.t("battlerTags:protectedLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-    return true;
-  }
-}
-
-/** Base class for `BattlerTag`s that block damaging moves but not status moves */
-export class DamageProtectedTag extends ProtectedTag {
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (attacker.getMoveCategory(pokemon, move) !== MoveCategory.STATUS) {
-      return super.apply(pokemon, simulated, attacker, move);
-    }
-    return false;
-  }
-}
-
-/**
- * `BattlerTag` class for moves that block damaging moves damage the enemy if the enemy's move makes contact
- * Used by {@linkcode MoveId.SPIKY_SHIELD}
- */
-export class ContactDamageProtectedTag extends ProtectedTag {
-  private damageRatio: number;
-
-  constructor(sourceMoveId: MoveId, damageRatio: number) {
-    super(sourceMoveId, BattlerTagType.SPIKY_SHIELD);
-
-    this.damageRatio = damageRatio;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.damageRatio = source.damageRatio;
-  }
-
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (!super.apply(pokemon, simulated, attacker, move)) {
-      return false;
-    }
-
-    if (!simulated && move.checkFlag(MoveFlags.MAKES_CONTACT, attacker, null)) {
-      if (!attacker.hasAbilityWithAttr(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE)) {
-        attacker.damageAndUpdate(toDmgValue(attacker.getMaxHp() * (1 / this.damageRatio)), HitResult.OTHER);
-      }
-    }
-    return true;
-  }
-}
-
-/**
- * `BattlerTag` class for moves that block damaging moves and lower enemy stats if the enemy's move makes contact
- * Used by {@linkcode MoveId.KINGS_SHIELD}, {@linkcode MoveId.OBSTRUCT}, {@linkcode MoveId.SILK_TRAP}
- */
-export class ContactStatStageChangeProtectedTag extends DamageProtectedTag {
-  private stat: BattleStat;
-  private levels: number;
-
-  constructor(sourceMoveId: MoveId, tagType: BattlerTagType, stat: BattleStat, levels: number) {
-    super(sourceMoveId, tagType);
-
-    this.stat = stat;
-    this.levels = levels;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.stat = source.stat;
-    this.levels = source.levels;
-  }
-
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (!super.apply(pokemon, simulated, attacker, move)) {
-      return false;
-    }
-
-    if (!simulated && move.checkFlag(MoveFlags.MAKES_CONTACT, attacker, null)) {
-      globalScene.unshiftPhase(new StatStageChangePhase(attacker.getBattlerIndex(), pokemon, [this.stat], this.levels));
-    }
-    return true;
-  }
-}
-
-export class ContactPoisonProtectedTag extends ProtectedTag {
-  constructor(sourceMoveId: MoveId) {
-    super(sourceMoveId, BattlerTagType.BANEFUL_BUNKER);
-  }
-
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (!super.apply(pokemon, simulated, attacker, move)) {
-      return false;
-    }
-
-    if (!simulated && move.checkFlag(MoveFlags.MAKES_CONTACT, attacker, null)) {
-      attacker.trySetStatus(StatusEffect.POISON, true, pokemon);
-    }
-    return true;
-  }
-}
-
-/**
- * `BattlerTag` class for moves that block damaging moves and burn the enemy if the enemy's move makes contact
- * Used by {@linkcode MoveId.BURNING_BULWARK}
- */
-export class ContactBurnProtectedTag extends DamageProtectedTag {
-  constructor(sourceMoveId: MoveId) {
-    super(sourceMoveId, BattlerTagType.BURNING_BULWARK);
-  }
-
-  override apply(pokemon: Pokemon, simulated: boolean, attacker: Pokemon, move: Move): boolean {
-    if (!super.apply(pokemon, simulated, attacker, move)) {
-      return false;
-    }
-
-    if (!simulated && move.checkFlag(MoveFlags.MAKES_CONTACT, attacker, null)) {
-      attacker.trySetStatus(StatusEffect.BURN, true);
-    }
-    return true;
-  }
-}
-
-/**
- * `BattlerTag` class for effects that cause the affected Pokemon to survive lethal attacks at 1 HP.
- * Used for {@link https://bulbapedia.bulbagarden.net/wiki/Endure_(move) | Endure} and
- * Endure Tokens.
- */
-export class EnduringTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, lapseType: BattlerTagLapseType, sourceMoveId: MoveId) {
-    super(tagType, lapseType, 0, sourceMoveId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:enduringOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.CUSTOM) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:enduringLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      return true;
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-/**
- * BattlerTag for the effects of {@link https://bulbapedia.bulbagarden.net/wiki/Magic_Coat_(move) | Magic Coat}.
- * Reflects status moves back at the attacker.
- * @extends BattlerTag
- * @see {@linkcode MovePhase.tryReflectMove}
- */
-export class MagicCoatTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.MAGIC_COAT, BattlerTagLapseType.TURN_END, 1);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    globalScene.queueMessage(
-      i18next.t("battlerTags:magicCoatOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override apply(
-    _pokemon: Pokemon,
-    simulated: boolean,
-    attacker: Pokemon,
-    move: Move,
-    reflected: BooleanHolder,
-  ): boolean {
-    if (!simulated) {
-      globalScene.queueMessage(this.getReflectionMessage(attacker, move));
-    }
-    reflected.value = true;
-    return true;
-  }
-
-  private getReflectionMessage(attacker: Pokemon, move: Move) {
-    // "{pokemonNameWithAffix}'s {moveName} was bounced back by Magic Coat!"
-    return i18next.t("battlerTags:magicCoatOnApply", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(attacker),
-      moveName: move.name,
-    });
-  }
-}
-
-export class SturdyTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.STURDY, BattlerTagLapseType.TURN_END, 0, sourceMoveId);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (lapseType === BattlerTagLapseType.CUSTOM) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:sturdyLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-      return true;
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-}
-
-export class PerishSongTag extends BattlerTag {
-  constructor(turnCount: number) {
-    super(BattlerTagType.PERISH_SONG, BattlerTagLapseType.TURN_END, turnCount, MoveId.PERISH_SONG, undefined, true);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.isBossImmune();
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:perishSongLapse", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          turnCount: this.turnCount,
-        }),
-      );
-    } else {
-      // The 2 here is just a number big enough to overcome the G-Max damage reduction
-      pokemon.damageAndUpdate(2 * pokemon.hp, HitResult.ONE_HIT_KO, false, true, true);
-    }
-
-    return ret;
-  }
-}
-
-/**
- * Applies the "Center of Attention" volatile status effect, the effect applied by Follow Me, Rage Powder, and Spotlight.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Center_of_attention | Center of Attention}
- */
-export class CenterOfAttentionTag extends BattlerTag {
-  public powder: boolean;
-
-  constructor(sourceMoveId: MoveId) {
-    super(BattlerTagType.CENTER_OF_ATTENTION, BattlerTagLapseType.TURN_END, 1, sourceMoveId);
-
-    this.powder = this.sourceMoveId === MoveId.RAGE_POWDER;
-  }
-
-  /** "Center of Attention" can't be added if an ally is already the Center of Attention. */
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.getField().find((p) => p.getTag(BattlerTagType.CENTER_OF_ATTENTION));
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:centerOfAttentionOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-export class AbilityBattlerTag extends BattlerTag {
-  public ability: Abilities;
-
-  constructor(tagType: BattlerTagType, ability: Abilities, lapseType: BattlerTagLapseType, turnCount: number) {
-    super(tagType, lapseType, turnCount);
-
-    this.ability = ability;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.ability = source.ability as Abilities;
-  }
-}
-
-/**
- * Tag used by Unburden to double speed
- * @extends AbilityBattlerTag
- */
-export class UnburdenTag extends AbilityBattlerTag {
-  constructor() {
-    super(BattlerTagType.UNBURDEN, Abilities.UNBURDEN, BattlerTagLapseType.CUSTOM, 1);
-  }
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-  }
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-  }
-}
-
-export class TruantTag extends AbilityBattlerTag {
-  constructor() {
-    super(BattlerTagType.TRUANT, Abilities.TRUANT, BattlerTagLapseType.MOVE, 1);
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (!pokemon.hasAbility(Abilities.TRUANT)) {
-      return super.lapse(pokemon, lapseType);
-    }
-    const passive = pokemon.getAbility().id !== Abilities.TRUANT;
-
-    const lastMove = pokemon.getLastXMoves().find(() => true);
-
-    if (lastMove && lastMove.move.id !== MoveId.NONE) {
-      globalScene.getCurrentPhase<MovePhase>()?.cancel();
-      globalScene.unshiftPhase(new ShowAbilityPhase(pokemon.id, passive));
-      globalScene.queueMessage(
-        i18next.t("battlerTags:truantLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-
-    return true;
-  }
-}
-
-export class SlowStartTag extends AbilityBattlerTag {
-  constructor() {
-    super(BattlerTagType.SLOW_START, Abilities.SLOW_START, BattlerTagLapseType.TURN_END, 5);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:slowStartOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      null,
-      false,
-      null,
-      true,
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    if (!pokemon.hasAbility(this.ability)) {
-      this.turnCount = 1;
-    }
-
-    return super.lapse(pokemon, lapseType);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:slowStartOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      null,
-      false,
-      null,
-    );
-  }
-}
-
-export class HighestStatBoostTag extends AbilityBattlerTag {
-  public stat: Stat;
-  public multiplier: number;
-
-  constructor(tagType: BattlerTagType, ability: Abilities) {
-    super(tagType, ability, BattlerTagLapseType.CUSTOM, 1);
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.stat = source.stat as Stat;
-    this.multiplier = source.multiplier;
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    let highestStat: EffectiveStat;
-    EFFECTIVE_STATS.map((s) => pokemon.getEffectiveStat(s)).reduce((highestValue: number, value: number, i: number) => {
-      if (value > highestValue) {
-        highestStat = EFFECTIVE_STATS[i];
-        return value;
-      }
-      return highestValue;
-    }, 0);
-
-    highestStat = highestStat!; // tell TS compiler it's defined!
-    this.stat = highestStat;
-
-    switch (this.stat) {
-      case Stat.SPD:
-        this.multiplier = 1.5;
-        break;
-      default:
-        this.multiplier = 1.3;
-        break;
-    }
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:highestStatBoostOnAdd", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        statName: i18next.t(getStatKey(highestStat)),
-      }),
-      null,
-      false,
-      null,
-      true,
-    );
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:highestStatBoostOnRemove", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        abilityName: allAbilities[this.ability].name,
-      }),
-    );
-  }
-}
-
-export class WeatherHighestStatBoostTag extends HighestStatBoostTag implements WeatherBattlerTag {
-  public weatherTypes: WeatherType[];
-
-  constructor(tagType: BattlerTagType, ability: Abilities, ...weatherTypes: WeatherType[]) {
-    super(tagType, ability);
-    this.weatherTypes = weatherTypes;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.weatherTypes = source.weatherTypes.map((w) => w as WeatherType);
-  }
-}
-
-export class TerrainHighestStatBoostTag extends HighestStatBoostTag implements TerrainBattlerTag {
-  public terrainTypes: TerrainType[];
-
-  constructor(tagType: BattlerTagType, ability: Abilities, ...terrainTypes: TerrainType[]) {
-    super(tagType, ability);
-    this.terrainTypes = terrainTypes;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.terrainTypes = source.terrainTypes.map((w) => w as TerrainType);
-  }
-}
-
-export class SemiInvulnerableTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, turnCount: number, sourceMoveId: MoveId) {
-    super(tagType, BattlerTagLapseType.MOVE_EFFECT, turnCount, sourceMoveId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    pokemon.setVisible(false);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    // Wait 2 frames before setting visible for battle animations that don't immediately show the sprite invisible
-    globalScene.tweens.addCounter({
-      duration: getFrameMs(2),
-      onComplete: () => pokemon.setVisible(true),
-    });
-  }
-}
-
-export class SkyDropTag extends BattlerTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.SKY_DROP, BattlerTagLapseType.CUSTOM, 1, MoveId.SKY_DROP, sourceId);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-    pokemon.setVisible(false);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    if (pokemon.id === this.sourceId) {
-      pokemon.removeTag(BattlerTagType.CHARGING);
-    }
-    // Wait 2 frames before setting visible for battle animations that don't immediately show the sprite invisible
-    globalScene.tweens.addCounter({
-      duration: getFrameMs(2),
-      onComplete: () => pokemon.setVisible(true),
-    });
-  }
-
-  /**
-   * Removes Sky Drop's effects from all Pokemon affected by this instance of Sky Drop.
-   */
-  public clearSkyDropEffects(): void {
-    globalScene.getField(true).forEach((pokemon) => {
-      if (pokemon?.getTag(BattlerTagType.SKY_DROP)?.sourceId === this.sourceId) {
-        // Cancel the Sky Drop user's next use of Sky Drop
-        if (this.sourceId === pokemon.id) {
-          globalScene.currentBattle.turnManager.tryRemoveCommand((tc) => tc.pokemon === pokemon);
-          if (
-            globalScene.tryRemovePhase((phase) => phase.is<MovePhase>(PhaseId.MOVE) && phase.pokemon.id === pokemon.id)
-          ) {
-            // Just in case we removed a queued `MovePhase`, queue the next `MovePhase`.
-            const { turnManager } = globalScene.currentBattle;
-            turnManager.scheduleNextValidCommand();
-          }
-          pokemon.getMoveQueue().shift();
-          pokemon.removeTag(BattlerTagType.CHARGING);
-        }
-        pokemon.removeTag(BattlerTagType.SKY_DROP);
-      }
-    });
-  }
-}
-
-export class TypeImmuneTag extends BattlerTag {
-  public immuneType: ElementalType;
-
-  constructor(tagType: BattlerTagType, sourceMoveId: MoveId, immuneType: ElementalType, length: number = 1) {
-    super(tagType, BattlerTagLapseType.TURN_END, length, sourceMoveId, undefined, true);
-
-    this.immuneType = immuneType;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.immuneType = source.immuneType as ElementalType;
-  }
-}
-
-/**
- * Battler Tag that lifts the affected Pokemon into the air and provides immunity to Ground type moves.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Magnet_Rise_(move) | MoveId.MAGNET_RISE}
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Telekinesis_(move) | MoveId.TELEKINESIS}
- */
-export class FloatingTag extends TypeImmuneTag {
-  constructor(tagType: BattlerTagType, sourceMoveId: MoveId, turnCount: number) {
-    super(tagType, sourceMoveId, ElementalType.GROUND, turnCount);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    if (this.sourceMoveId === MoveId.MAGNET_RISE) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:magnetRisenOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-    if (this.sourceMoveId === MoveId.MAGNET_RISE) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:magnetRisenOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-  }
-}
-
-export class TypeBoostTag extends BattlerTag {
-  public boostedType: ElementalType;
-  public boostValue: number;
-  public oneUse: boolean;
-
-  constructor(
-    tagType: BattlerTagType,
-    sourceMoveId: MoveId,
-    boostedType: ElementalType,
-    boostValue: number,
-    oneUse: boolean,
-  ) {
-    super(tagType, BattlerTagLapseType.TURN_END, 1, sourceMoveId);
-
-    this.boostedType = boostedType;
-    this.boostValue = boostValue;
-    this.oneUse = oneUse;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.boostedType = source.boostedType as ElementalType;
-    this.boostValue = source.boostValue;
-    this.oneUse = source.oneUse;
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    return lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-  }
-
-  override isTypeBoostTag(): this is this {
-    return true;
-  }
-}
-
-/**
- * Tag to denote a nonstackable boost to crit rate. Granted by:
- * Focus Energy (+2), Dragon Cheer (+2 if dragon, +1 otherwise),
- * and Lansat Berry (+2)
- * @extends BattlerTag
- */
-export class CritBoostTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, sourceMoveId: MoveId) {
-    super(tagType, BattlerTagLapseType.CUSTOM, 1, sourceMoveId, undefined, true);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:critBoostOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    return lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:critBoostOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-/**
- * A stackable instance of crit boost granted by G-Max Chi Strike
- * @extends BattlerTag
- */
-export class CritBoostStackableTag extends BattlerTag {
-  public stackCount: number = 0;
-
-  constructor() {
-    super(BattlerTagType.CRIT_BOOST_STACKABLE, BattlerTagLapseType.CUSTOM, 1, MoveId.G_MAX_CHI_STRIKE, undefined);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    this.stackCount += 1;
-    // This actually does not have any messages in the mainline games
-    globalScene.queueMessage(
-      i18next.t("battlerTags:critBoostOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    this.onAdd(pokemon);
-  }
-
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.stackCount = source.stackCount ?? 0;
-  }
-}
-
-/**
- * Tag for the effects of Dragon Cheer, which boosts the critical hit ratio of the user's allies.
- * @extends {CritBoostTag}
- */
-export class DragonCheerTag extends CritBoostTag {
-  /** The types of the user's ally when the tag is added */
-  public typesOnAdd: ElementalType[];
-
-  constructor() {
-    super(BattlerTagType.CRIT_BOOST, MoveId.DRAGON_CHEER);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    this.typesOnAdd = pokemon.getTypes(true);
-  }
-}
-
-export class SaltCuredTag extends BattlerTag {
-  private sourceIndex: number;
-
-  constructor(sourceId: number) {
-    super(BattlerTagType.SALT_CURED, BattlerTagLapseType.TURN_END, 1, MoveId.SALT_CURE, sourceId);
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.sourceIndex = source.sourceIndex;
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battlerTags:saltCuredOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-    this.sourceIndex = globalScene.getPokemonById(this.sourceId!)!.getBattlerIndex(); // TODO: are those bangs correct?
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.unshiftPhase(
-        new CommonAnimPhase(pokemon.getBattlerIndex(), pokemon.getBattlerIndex(), CommonAnim.SALT_CURE),
-      );
-
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-
-      if (!cancelled.value) {
-        const pokemonSteelOrWater = pokemon.isOfType(ElementalType.STEEL) || pokemon.isOfType(ElementalType.WATER);
-        pokemon.damageAndUpdate(toDmgValue(pokemonSteelOrWater ? pokemon.getMaxHp() / 4 : pokemon.getMaxHp() / 8));
-
-        globalScene.queueMessage(
-          i18next.t("battlerTags:saltCuredLapse", {
-            pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-            moveName: this.getMoveName(),
-          }),
-        );
-      }
-    }
-
-    return ret;
-  }
-}
-
-export class CursedTag extends BattlerTag {
-  private sourceIndex: number;
-
-  constructor(sourceId: number) {
-    super(BattlerTagType.CURSED, BattlerTagLapseType.TURN_END, 1, MoveId.CURSE, sourceId, true);
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.sourceIndex = source.sourceIndex;
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-    this.sourceIndex = globalScene.getPokemonById(this.sourceId!)!.getBattlerIndex(); // TODO: are those bangs correct?
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = lapseType !== BattlerTagLapseType.CUSTOM || super.lapse(pokemon, lapseType);
-
-    if (ret) {
-      globalScene.unshiftPhase(
-        new CommonAnimPhase(pokemon.getBattlerIndex(), pokemon.getBattlerIndex(), CommonAnim.SALT_CURE),
-      );
-
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, pokemon, false, cancelled);
-
-      if (!cancelled.value) {
-        pokemon.damageAndUpdate(toDmgValue(pokemon.getMaxHp() / 4));
-        globalScene.queueMessage(
-          i18next.t("battlerTags:cursedLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        );
-      }
-    }
-
-    return ret;
-  }
-}
-/**
- * Battler tag for attacks that remove a type post use.
- */
-export class RemovedTypeTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, lapseType: BattlerTagLapseType, sourceMoveId: MoveId) {
-    super(tagType, lapseType, 1, sourceMoveId);
-  }
-}
-
-/**
- * Battler tag for effects that ground the source, allowing Ground-type moves to hit them.
- * @description `IGNORE_FLYING`: Persistent grounding effects (i.e. from Smack Down and Thousand Waves)
- */
-export class GroundedTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, lapseType: BattlerTagLapseType, sourceMoveId: MoveId) {
-    super(tagType, lapseType, 1, sourceMoveId);
-  }
-
-  /**
-   * Smack Down and Thousand Arrows have special messages when knocking an ungrounded Pokemon down
-   * @param pokemon the Pokemon being grounded
-   */
-  override onAdd(pokemon: Pokemon) {
-    const isSmackDownOrThousandArrows = [MoveId.SMACK_DOWN, MoveId.THOUSAND_ARROWS].includes(this.sourceMoveId);
-    const wasNotGrounded =
-      pokemon.isOfType(ElementalType.FLYING, true, true)
-      || pokemon.hasAbility(Abilities.LEVITATE)
-      || pokemon.getTag(BattlerTagType.FLOATING)
-      || pokemon.getTag(...SemiInvulnerableBattlerTagTypes);
-
-    if (isSmackDownOrThousandArrows && wasNotGrounded) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:groundedSmackDown", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      );
-    }
-  }
-}
-
-/**
- * @description `ROOSTED`: Tag for temporary grounding if only source of ungrounding is flying and pokemon uses Roost.
- * Roost removes flying type from a pokemon for a single turn.
- */
-
-export class RoostedTag extends BattlerTag {
-  private isBaseFlying: boolean;
-  private isBasePureFlying: boolean;
-
-  constructor() {
-    super(BattlerTagType.ROOSTED, BattlerTagLapseType.TURN_END, 1, MoveId.ROOST);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    const currentTypes = pokemon.getTypes();
-    const baseTypes = pokemon.getTypes(false, false, true);
-
-    const forestsCurseApplied: boolean =
-      currentTypes.includes(ElementalType.GRASS) && !baseTypes.includes(ElementalType.GRASS);
-    const trickOrTreatApplied: boolean =
-      currentTypes.includes(ElementalType.GHOST) && !baseTypes.includes(ElementalType.GHOST);
-
-    if (this.isBaseFlying) {
-      let modifiedTypes: ElementalType[] = [];
-      if (this.isBasePureFlying) {
-        if (forestsCurseApplied || trickOrTreatApplied) {
-          modifiedTypes = currentTypes.filter((type) => type !== ElementalType.NORMAL);
-          modifiedTypes.push(ElementalType.FLYING);
-        } else {
-          modifiedTypes = [ElementalType.FLYING];
-        }
-      } else {
-        modifiedTypes = [...currentTypes];
-        modifiedTypes.push(ElementalType.FLYING);
-      }
-      pokemon.summonData.types = modifiedTypes;
-      pokemon.updateInfo();
-    }
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    const currentTypes = pokemon.getTypes();
-    const baseTypes = pokemon.getTypes(false, false, true);
-
-    const isOriginallyDualType = baseTypes.length === 2;
-    const isCurrentlyDualType = currentTypes.length === 2;
-    this.isBaseFlying = baseTypes.includes(ElementalType.FLYING);
-    this.isBasePureFlying = baseTypes[0] === ElementalType.FLYING && baseTypes.length === 1;
-
-    if (this.isBaseFlying) {
-      let modifiedTypes: ElementalType[];
-      if (this.isBasePureFlying && !isCurrentlyDualType) {
-        modifiedTypes = [ElementalType.NORMAL];
-      } else {
-        if (!!pokemon.getTag(...RemoveTypeBattlerTagTypes) && isOriginallyDualType && !isCurrentlyDualType) {
-          modifiedTypes = [ElementalType.UNKNOWN];
-        } else {
-          modifiedTypes = currentTypes.filter((type) => type !== ElementalType.FLYING);
-        }
-      }
-      pokemon.summonData.types = modifiedTypes;
-      pokemon.updateInfo();
-    }
-  }
-}
-
-/** Common attributes of form change abilities that block damage */
-export class FormBlockDamageTag extends BattlerTag {
-  constructor(tagType: BattlerTagType) {
-    super(tagType, BattlerTagLapseType.CUSTOM, 1);
-  }
-
-  /**
-   * Determines if the tag can be added to the Pokémon.
-   * @param pokemon The Pokémon to which the tag might be added.
-   * @returns True if the tag can be added, false otherwise.
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    return pokemon.formIndex === 0;
-  }
-
-  /**
-   * Applies the tag to the Pokémon.
-   * Triggers a form change if the Pokémon is not in its defense form.
-   * @param pokemon The Pokémon to which the tag is added.
-   */
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-
-    if (pokemon.formIndex !== 0) {
-      globalScene.triggerPokemonFormChange(pokemon, SpeciesFormChangeManualTrigger);
-    }
-  }
-
-  /**
-   * Removes the tag from the Pokémon.
-   * Triggers a form change when the tag is removed.
-   * @param pokemon The Pokémon from which the tag is removed.
-   */
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.triggerPokemonFormChange(pokemon, SpeciesFormChangeManualTrigger);
-  }
-}
-/** Provides the additional weather-based effects of the Ice Face ability */
-export class IceFaceBlockDamageTag extends FormBlockDamageTag {
-  constructor(tagType: BattlerTagType) {
-    super(tagType);
-  }
-
-  /**
-   * Determines if the tag can be added to the Pokémon.
-   * @param pokemon The Pokémon to which the tag might be added.
-   * @returns True if the tag can be added, false otherwise.
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    return super.canAdd(pokemon) || globalScene.arena.hasWeather([WeatherType.HAIL, WeatherType.SNOW]);
-  }
-}
-
-/**
- * Battler tag indicating a Tatsugiri with {@link https://bulbapedia.bulbagarden.net/wiki/Commander_(Ability) | Commander}
- * has entered the tagged Pokemon's mouth.
- */
-export class CommandedTag extends BattlerTag {
-  private _tatsugiriFormKey: string;
-
-  constructor(sourceId: number) {
-    super(BattlerTagType.COMMANDED, BattlerTagLapseType.CUSTOM, 0, MoveId.NONE, sourceId);
-  }
-
-  public get tatsugiriFormKey(): string {
-    return this._tatsugiriFormKey;
-  }
-
-  /** Caches the Tatsugiri's form key and sharply boosts the tagged Pokemon's stats */
-  override onAdd(pokemon: Pokemon): void {
-    this._tatsugiriFormKey = this.getSourcePokemon()?.getFormKey() ?? "curly";
-    globalScene.unshiftPhase(
-      new StatStageChangePhase(
-        pokemon.getBattlerIndex(),
-        pokemon,
-        [Stat.ATK, Stat.DEF, Stat.SPATK, Stat.SPDEF, Stat.SPD],
-        2,
-      ),
-    );
-  }
-
-  /** Triggers an {@linkcode PokemonAnimType | animation} of the tagged Pokemon "spitting out" Tatsugiri */
-  override onRemove(pokemon: Pokemon): void {
-    if (this.getSourcePokemon()?.isActive(true)) {
-      globalScene.triggerPokemonBattleAnim(pokemon, PokemonAnimType.COMMANDER_REMOVE);
-    }
-  }
-
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this._tatsugiriFormKey = source._tatsugiriFormKey;
-  }
-}
-
-/**
- * Battler tag enabling the Stockpile mechanic. This tag handles:
- * - Stack tracking, including max limit enforcement (which is replicated in Stockpile for redundancy).
- *
- * - Stat changes on adding a stack. Adding a stockpile stack attempts to raise the pokemon's DEF and SPDEF by +1.
- *
- * - Stat changes on removal of (all) stacks.
- *   - Removing stacks decreases DEF and SPDEF, independently, by one stage for each stack that successfully changed
- *     the stat when added.
- */
-export class StockpilingTag extends BattlerTag {
-  public stockpiledCount: number = 0;
-  public statChangeCounts: { [Stat.DEF]: number; [Stat.SPDEF]: number } = {
-    [Stat.DEF]: 0,
-    [Stat.SPDEF]: 0,
-  };
-
-  constructor(sourceMoveId: MoveId = MoveId.NONE) {
-    super(BattlerTagType.STOCKPILING, BattlerTagLapseType.CUSTOM, 1, sourceMoveId);
-  }
-
-  private onStatStagesChanged: StatStageChangeCallback = (statsChanged, statChanges) => {
-    const defChange = statChanges[statsChanged.indexOf(Stat.DEF)] ?? 0;
-    const spDefChange = statChanges[statsChanged.indexOf(Stat.SPDEF)] ?? 0;
-
-    if (defChange) {
-      this.statChangeCounts[Stat.DEF]++;
-    }
-    if (spDefChange) {
-      this.statChangeCounts[Stat.SPDEF]++;
-    }
-  };
-
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.stockpiledCount = source.stockpiledCount ?? 0;
-    this.statChangeCounts = {
-      [Stat.DEF]: source.statChangeCounts?.[Stat.DEF] ?? 0,
-      [Stat.SPDEF]: source.statChangeCounts?.[Stat.SPDEF] ?? 0,
-    };
-  }
-
-  /**
-   * Adds a stockpile stack to a pokemon, up to a maximum of 3 stacks. Note that onOverlap defers to this method.
-   *
-   * If a stack is added, a message is displayed and the pokemon's DEF and SPDEF are increased by 1.
-   * For each stat, an internal counter is incremented (by 1) if the stat was successfully changed.
-   */
-  override onAdd(pokemon: Pokemon): void {
-    if (this.stockpiledCount < 3) {
-      this.stockpiledCount++;
-
-      globalScene.queueMessage(
-        i18next.t("battlerTags:stockpilingOnAdd", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-          stockpiledCount: this.stockpiledCount,
-        }),
-      );
-
-      // Attempt to increase DEF and SPDEF by one stage, keeping track of successful changes.
-      globalScene.unshiftPhase(
-        new StatStageChangePhase(pokemon.getBattlerIndex(), pokemon, [Stat.SPDEF, Stat.DEF], 1, {
-          onChange: this.onStatStagesChanged,
-        }),
-      );
-    }
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    this.onAdd(pokemon);
-  }
-
-  /**
-   * Removing the tag removes all stacks, and the pokemon's DEF and SPDEF are decreased by
-   * one stage for each stack which had successfully changed that particular stat during onAdd.
-   */
-  override onRemove(pokemon: Pokemon): void {
-    const defChange = this.statChangeCounts[Stat.DEF];
-    const spDefChange = this.statChangeCounts[Stat.SPDEF];
-
-    if (defChange) {
-      globalScene.unshiftPhase(new StatStageChangePhase(pokemon.getBattlerIndex(), pokemon, [Stat.DEF], -defChange));
-    }
-
-    if (spDefChange) {
-      globalScene.unshiftPhase(
-        new StatStageChangePhase(pokemon.getBattlerIndex(), pokemon, [Stat.SPDEF], -spDefChange),
-      );
-    }
-  }
-}
-
-/**
- * Battler tag for Gulp Missile used by Cramorant.
- * @extends BattlerTag
- */
-export class GulpMissileTag extends BattlerTag {
-  constructor(tagType: BattlerTagType, sourceMoveId: MoveId) {
-    super(tagType, BattlerTagLapseType.HIT, 0, sourceMoveId);
-  }
-
-  override lapse(pokemon: Pokemon, _lapseType: BattlerTagLapseType): boolean {
-    if (pokemon.getTag(BattlerTagType.UNDERWATER)) {
-      return true;
-    }
-
-    const moveEffectPhase = globalScene.getCurrentPhase();
-    if (moveEffectPhase?.is<MoveEffectPhase>(PhaseId.MOVE_EFFECT)) {
-      const attacker = moveEffectPhase.getUserPokemon();
-
-      if (!attacker) {
-        return false;
-      }
-
-      if (moveEffectPhase.move.getMove().hitsSubstitute(attacker, pokemon)) {
-        return true;
-      }
-
-      const cancelled = new BooleanHolder(false);
-      applyAbAttrs(AbAttrFlag.BLOCK_NON_DIRECT_DAMAGE, attacker, false, cancelled);
-
-      if (!cancelled.value) {
-        attacker.damageAndUpdate(Math.max(1, Math.floor(attacker.getMaxHp() / 4)), HitResult.OTHER);
-      }
-
-      if (this.tagType === BattlerTagType.GULP_MISSILE_ARROKUDA) {
-        globalScene.unshiftPhase(new StatStageChangePhase(attacker.getBattlerIndex(), pokemon, [Stat.DEF], -1));
-      } else {
-        attacker.trySetStatus(StatusEffect.PARALYSIS, true, pokemon);
-      }
-    }
-    return false;
-  }
-
-  /**
-   * Gulp Missile's initial form changes are triggered by using Surf and Dive.
-   * @param pokemon The Pokemon with Gulp Missile ability.
-   * @returns Whether the BattlerTag can be added.
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    const isSurfOrDive = [MoveId.SURF, MoveId.DIVE].includes(this.sourceMoveId);
-    const isNormalForm = pokemon.formIndex === 0 && !pokemon.getTag(...GulpMissileBattlerTagTypes);
-    const isCramorant = pokemon.species.speciesId === Species.CRAMORANT;
-
-    return isSurfOrDive && isNormalForm && isCramorant;
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-    globalScene.triggerPokemonFormChange(pokemon, SpeciesFormChangeManualTrigger);
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-    globalScene.triggerPokemonFormChange(pokemon, SpeciesFormChangeManualTrigger);
-  }
-}
-
-/**
- * Tag that makes the target drop all of it type immunities
- * and all accuracy checks ignore its evasiveness stat.
- *
- * Applied by moves: {@linkcode MoveId.ODOR_SLEUTH | Odor Sleuth},
- * {@linkcode MoveId.MIRACLE_EYE | Miracle Eye} and {@linkcode MoveId.FORESIGHT | Foresight}.
- *
- * @extends BattlerTag
- * @see {@linkcode ignoreImmunity}
- */
-export class ExposedTag extends BattlerTag {
-  private defenderType: ElementalType;
-  private allowedTypes: ElementalType[];
-
-  constructor(
-    tagType: BattlerTagType,
-    sourceMoveId: MoveId,
-    defenderType: ElementalType,
-    allowedTypes: ElementalType[],
-  ) {
-    super(tagType, BattlerTagLapseType.CUSTOM, 1, sourceMoveId);
-    this.defenderType = defenderType;
-    this.allowedTypes = allowedTypes;
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.defenderType = source.defenderType as ElementalType;
-    this.allowedTypes = source.allowedTypes as ElementalType[];
-  }
-
-  /**
-   * @param types {@linkcode ElementalType} of the defending Pokemon
-   * @param moveType {@linkcode ElementalType} of the move targetting it
-   * @returns `true` if the move should be allowed to target the defender.
-   */
-  ignoreImmunity(type: ElementalType, moveType: ElementalType): boolean {
-    return type === this.defenderType && this.allowedTypes.includes(moveType);
-  }
-}
-
-/**
- * Tag that prevents HP recovery from held items and move effects. It also blocks the usage of recovery moves.
- * Applied by moves:  {@linkcode MoveId.HEAL_BLOCK | Heal Block (5 turns)}, {@linkcode MoveId.PSYCHIC_NOISE | Psychic Noise (2 turns)}
- *
- * @extends MoveRestrictionBattlerTag
- */
-export class HealBlockTag extends MoveRestrictionBattlerTag {
-  constructor(turnCount: number, sourceMoveId: MoveId) {
-    super(
-      BattlerTagType.HEAL_BLOCK,
-      [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.TURN_END],
-      turnCount,
-      sourceMoveId,
-    );
-  }
-
-  onActivation(pokemon: Pokemon): string {
-    return i18next.t("battle:battlerTagsHealBlock", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-
-  /**
-   * Checks if a move is disabled under Heal Block
-   * @param moveId {@linkcode MoveId} the move ID
-   * @returns `true` if the move has a TRIAGE_MOVE flag
-   */
-  override isMoveRestricted(moveId: MoveId): boolean {
-    if (allMoves.get(moveId).hasFlag(MoveFlags.TRIAGE_MOVE)) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Checks if a move is disabled under Heal Block because of its choice of target
-   * Implemented b/c of Pollen Puff
-   * @param moveId {@linkcode MoveId} the move ID
-   * @param user {@linkcode Pokemon} the move user
-   * @param target {@linkcode Pokemon} the target of the move
-   * @returns `true` if the move cannot be used because the target is an ally
-   */
-  override isMoveTargetRestricted(moveId: MoveId, user: Pokemon, target: Pokemon) {
-    const moveCategory = new NumberHolder(allMoves.get(moveId).category);
-    applyMoveAttrs(StatusCategoryOnAllyAttr, user, target, allMoves.get(moveId), moveCategory);
-    if (allMoves.get(moveId).hasAttr(HealOnAllyAttr) && moveCategory.value === MoveCategory.STATUS) {
-      return true;
-    }
-    return false;
-  }
-
-  /**
-   * Uses its own unique getSelectionDeniedText() message
-   */
-  override getSelectionDeniedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledHealBlock", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-      healBlockName: allMoves.get(MoveId.HEAL_BLOCK).name,
-    });
-  }
-
-  /**
-   * @override
-   * @param pokemon {@linkcode Pokemon} attempting to use the restricted move
-   * @param moveId {@linkcode MoveId} ID of the move being interrupted
-   * @returns text to display when the move is interrupted
-   */
-  override getInterruptedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledHealBlock", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-      healBlockName: allMoves.get(MoveId.HEAL_BLOCK).name,
-    });
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-
-    globalScene.queueMessage(
-      i18next.t("battle:battlerTagsHealBlockOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      null,
-      false,
-      null,
-    );
-  }
-}
-
-/**
- * Tag that doubles the type effectiveness of Fire-type moves.
- * @extends BattlerTag
- */
-export class TarShotTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.TAR_SHOT, BattlerTagLapseType.CUSTOM, 0);
-  }
-
-  /**
-   * If the Pokemon is terastallized, the tag cannot be added.
-   * @param pokemon the {@linkcode Pokemon} to which the tag is added
-   * @returns whether the tag is applied
-   */
-  override canAdd(pokemon: Pokemon): boolean {
-    return !pokemon.isTerastallized();
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    globalScene.queueMessage(
-      i18next.t("battlerTags:tarShotOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-/**
- * Battler Tag implementing the type-changing effect of {@link https://bulbapedia.bulbagarden.net/wiki/Electrify_(move) | Electrify}.
- * While this tag is in effect, the afflicted Pokemon's moves are changed to Electric type.
- */
-export class ElectrifiedTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.ELECTRIFIED, BattlerTagLapseType.TURN_END, 1, MoveId.ELECTRIFY);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    // "{pokemonNameWithAffix}'s moves have been electrified!"
-    globalScene.queueMessage(
-      i18next.t("battlerTags:electrifiedOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-/**
- * Battler Tag that keeps track of how many times the user has Autotomized
- * Each count of Autotomization reduces the weight by 100kg
- */
-export class AutotomizedTag extends BattlerTag {
-  public autotomizeCount: number = 0;
-  constructor(sourceMoveId: MoveId = MoveId.AUTOTOMIZE) {
-    super(BattlerTagType.AUTOTOMIZED, BattlerTagLapseType.CUSTOM, 1, sourceMoveId);
-  }
-
-  /**
-   * Adds an autotomize count to the Pokemon. Each stack reduces weight by 100kg
-   * If the Pokemon is over 0.1kg it also displays a message.
-   * @param pokemon The Pokemon that is being autotomized
-   */
-  override onAdd(pokemon: Pokemon): void {
-    const minWeight = 0.1;
-    if (pokemon.getWeight() > minWeight) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:autotomizeOnAdd", {
-          pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-        }),
-      );
-    }
-    this.autotomizeCount += 1;
-  }
-
-  override onOverlap(pokemon: Pokemon): void {
-    this.onAdd(pokemon);
-  }
-}
-
-/**
- * Tag implementing the {@link https://bulbapedia.bulbagarden.net/wiki/Substitute_(doll)#Effect | Substitute Doll} effect,
- * for use with the moves Substitute and Shed Tail. Pokemon with this tag deflect most forms of received attack damage
- * onto the tag. This tag also grants immunity to most Status moves and several move effects.
- */
-export class SubstituteTag extends BattlerTag {
-  /** The substitute's remaining HP. If HP is depleted, the Substitute fades. */
-  public hp: number;
-  /** A reference to the sprite representing the Substitute doll */
-  public sprite: Phaser.GameObjects.Sprite;
-  /** Is the source Pokemon "in focus," i.e. is it fully visible on the field? */
-  public sourceInFocus: boolean;
-
-  constructor(sourceMoveId: MoveId, sourceId: number) {
-    super(
-      BattlerTagType.SUBSTITUTE,
-      [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.AFTER_MOVE, BattlerTagLapseType.HIT],
-      0,
-      sourceMoveId,
-      sourceId,
-      true,
-    );
-  }
-
-  /** Sets the Substitute's HP and queues an on-add battle animation that initializes the Substitute's sprite. */
-  override onAdd(pokemon: Pokemon): void {
-    this.hp = Math.floor(globalScene.getPokemonById(this.sourceId!)!.getMaxHp() / 4);
-    this.sourceInFocus = false;
-
-    // Queue battle animation and message
-    globalScene.triggerPokemonBattleAnim(pokemon, PokemonAnimType.SUBSTITUTE_ADD);
-    if (this.sourceMoveId === MoveId.SHED_TAIL) {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:shedTailOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        1500,
-      );
-    } else {
-      globalScene.queueMessage(
-        i18next.t("battlerTags:substituteOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        1500,
-      );
-    }
-
-    // Remove any binding effects from the user
-    pokemon.findAndRemoveTags((tag) => tag instanceof DamagingTrapTag);
-  }
-
-  /** Queues an on-remove battle animation that removes the Substitute's sprite. */
-  override onRemove(pokemon: Pokemon): void {
-    // Only play the animation if the cause of removal isn't from the source's own move
-    if (!this.sourceInFocus) {
-      globalScene.triggerPokemonBattleAnim(pokemon, PokemonAnimType.SUBSTITUTE_REMOVE, [this.sprite]);
-    } else {
-      this.sprite.destroy();
-    }
-    globalScene.queueMessage(
-      i18next.t("battlerTags:substituteOnRemove", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    switch (lapseType) {
-      case BattlerTagLapseType.PRE_MOVE:
-        this.onPreMove(pokemon);
-        break;
-      case BattlerTagLapseType.AFTER_MOVE:
-        this.onAfterMove(pokemon);
-        break;
-      case BattlerTagLapseType.HIT:
-        this.onHit(pokemon);
-        break;
-    }
-    return lapseType !== BattlerTagLapseType.CUSTOM; // only remove this tag on custom lapse
-  }
-
-  /** Triggers an animation that brings the Pokemon into focus before it uses a move */
-  onPreMove(pokemon: Pokemon): void {
-    globalScene.triggerPokemonBattleAnim(pokemon, PokemonAnimType.SUBSTITUTE_PRE_MOVE, [this.sprite]);
-    this.sourceInFocus = true;
-  }
-
-  /** Triggers an animation that brings the Pokemon out of focus after it uses a move */
-  onAfterMove(pokemon: Pokemon): void {
-    globalScene.triggerPokemonBattleAnim(pokemon, PokemonAnimType.SUBSTITUTE_POST_MOVE, [this.sprite]);
-    this.sourceInFocus = false;
-  }
-
-  /** If the Substitute redirects damage, queue a message to indicate it. */
-  onHit(pokemon: Pokemon): void {
-    const moveEffectPhase = globalScene.getCurrentPhase();
-    if (moveEffectPhase?.is<MoveEffectPhase>(PhaseId.MOVE_EFFECT)) {
-      const attacker = moveEffectPhase.getUserPokemon();
-      if (!attacker) {
-        return;
-      }
-      const move = moveEffectPhase.move.getMove();
-      const firstHit = attacker.turnData.hitCount === attacker.turnData.hitsLeft;
-
-      if (firstHit && move.hitsSubstitute(attacker, pokemon)) {
-        globalScene.queueMessage(
-          i18next.t("battlerTags:substituteOnHit", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-        );
-      }
-    }
-  }
-
-  /**
-   * When given a battler tag or json representing one, load the data for it.
-   * @param source A battler tag
-   */
-  override loadTag(source: BattlerTag | any): void {
-    super.loadTag(source);
-    this.hp = source.hp;
-  }
-}
-
-/**
- * Tag that adds extra post-summon effects to a battle for a specific Pokemon.
- * These post-summon effects are performed through {@linkcode Pokemon.mysteryEncounterBattleEffects},
- * and can be used to unshift special phases, etc.
- * Currently used only in MysteryEncounters to provide start of fight stat buffs.
- */
-export class MysteryEncounterPostSummonTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.MYSTERY_ENCOUNTER_POST_SUMMON, BattlerTagLapseType.CUSTOM, 1);
-  }
-
-  /** Event when tag is added */
-  override onAdd(pokemon: Pokemon): void {
-    super.onAdd(pokemon);
-  }
-
-  /** Performs post-summon effects through {@linkcode Pokemon.mysteryEncounterBattleEffects} */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType): boolean {
-    const ret = super.lapse(pokemon, lapseType);
-
-    if (lapseType === BattlerTagLapseType.CUSTOM) {
-      if (pokemon.mysteryEncounterBattleEffects) {
-        pokemon.mysteryEncounterBattleEffects(pokemon);
-      }
-    }
-
-    return ret;
-  }
-
-  /** Event when tag is removed */
-  override onRemove(pokemon: Pokemon): void {
-    super.onRemove(pokemon);
-  }
-
-  override isMysteryEncounterPostSummonTag(): this is this {
-    return true;
-  }
-}
-
-/**
- * Battle Tag that applies the move Torment to the target Pokemon
- * Torment restricts the use of moves twice in a row.
- * The tag is only removed if the target leaves the battle.
- * Torment does not interrupt the move if the move is performed consecutively in the same turn and right after Torment is applied
- */
-export class TormentTag extends MoveRestrictionBattlerTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.TORMENT, BattlerTagLapseType.AFTER_MOVE, 1, MoveId.TORMENT, sourceId);
-  }
-
-  /**
-   * Adds the battler tag to the target Pokemon and defines the private class variable 'target'
-   * 'Target' is used to track the Pokemon's current status
-   * @param pokemon the Pokemon tormented
-   */
-  override onAdd(pokemon: Pokemon) {
-    super.onAdd(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:tormentOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      1500,
-    );
-  }
-
-  /**
-   * Torment only ends when the affected Pokemon leaves the battle field
-   * @param pokemon the Pokemon under the effects of Torment
-   * @param _tagType
-   * @returns `true` if still present | `false` if not
-   */
-  override lapse(pokemon: Pokemon, _tagType: BattlerTagLapseType): boolean {
-    return pokemon.isActive(true);
-  }
-
-  /**
-   * This checks if the current move used is identical to the last used move with a {@linkcode MoveResult} of `SUCCESS`/`MISS`
-   * @param moveId the move under investigation
-   * @returns `true` if there is valid consecutive usage | `false` if the moves are different from each other
-   */
-  public override isMoveRestricted(moveId: MoveId, user: Pokemon): boolean {
-    if (!user) {
-      return false;
-    }
-    const lastMoveTurn = user.getLastXMoves(1)[0];
-    if (!lastMoveTurn) {
-      return false;
-    }
-
-    const validLastMoveResult = lastMoveTurn.result === MoveResult.SUCCESS || lastMoveTurn.result === MoveResult.MISS;
-    if (
-      lastMoveTurn.move.id === moveId
-      && validLastMoveResult
-      && lastMoveTurn.move.id !== MoveId.STRUGGLE
-      && !user.getTag(...MoveLockTagTypes)
-    ) {
-      return true;
-    }
-    return false;
-  }
-
-  override getSelectionDeniedText(pokemon: Pokemon, _moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledTorment", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) });
-  }
-
-  override getInterruptedText(_pokemon: Pokemon, _moveId: MoveId): string {
-    return "";
-  }
-}
-
-/**
- * BattlerTag that applies the effects of Taunt to the target Pokemon
- * Taunt restricts the use of status moves.
- * The tag is removed after 4 turns.
- */
-export class TauntTag extends MoveRestrictionBattlerTag {
-  constructor() {
-    super(BattlerTagType.TAUNT, [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.AFTER_MOVE], 4, MoveId.TAUNT);
-  }
-
-  override onAdd(pokemon: Pokemon) {
-    super.onAdd(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:tauntOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-      1500,
-    );
-  }
-
-  /**
-   * Checks if a move is a status move and determines its restriction status on that basis
-   * @param moveId the move under investigation
-   * @returns `true` if the move is a status move
-   */
-  override isMoveRestricted(moveId: MoveId): boolean {
-    return allMoves.get(moveId).category === MoveCategory.STATUS;
-  }
-
-  override getSelectionDeniedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledTaunt", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-    });
-  }
-
-  override getInterruptedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledTaunt", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-    });
-  }
-}
-
-/**
- * BattlerTag representing the move-disabling effect of
- * {@link https://bulbapedia.bulbagarden.net/wiki/Imprison_(move) | Imprison}.
- * Disables all opposing Pokemon's moves that are also found in the tag owner's moveset.
- * @extends BattlerTag
- * @implements RestrictingBattlerTag
- */
-export class ImprisoningTag extends BattlerTag implements RestrictingBattlerTag {
-  constructor() {
-    super(BattlerTagType.IMPRISONING, BattlerTagLapseType.CUSTOM, 1);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    globalScene.queueMessage(
-      i18next.t("battlerTags:imprisonOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Prevents opposing Pokemon from using moves that match with
-   * the tag owner's moveset.
-   * @param pokemon The {@linkcode Pokemon} with this tag
-   * @param simulated If `true`, suppresses the message triggered by an interruption
-   * @param actingPokemon The {@linkcode Pokemon} attempting to select or use a move
-   * @param moveId The {@linkcode MoveId} for the move being selected or used
-   * @returns `true` if the given move is disabled by this tag
-   */
-  override apply(pokemon: Pokemon, simulated: boolean, actingPokemon: Pokemon, moveId: MoveId): boolean {
-    if (pokemon.getMoveset().some((mv) => mv.moveId === moveId)) {
-      if (!simulated) {
-        globalScene.queueMessage(this.getInterruptedText(actingPokemon, moveId));
-      }
-      return true;
-    }
-    return false;
-  }
-
-  public getInterruptedText(pokemon: Pokemon, moveId: MoveId): string {
-    return i18next.t("battle:moveDisabledImprison", {
-      pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      moveName: allMoves.get(moveId).name,
-    });
-  }
-
-  public getSelectionDeniedText(pokemon: Pokemon, moveId: MoveId): string {
-    return this.getInterruptedText(pokemon, moveId);
-  }
-}
-
-/**
- * Battler Tag that applies the effects of Syrup Bomb to the target Pokemon.
- * For three turns, starting from the turn of hit, at the end of each turn, the target Pokemon's speed will decrease by 1.
- * The tag can also expire by taking the target Pokemon off the field, or the Pokemon that originally used the move.
- */
-export class SyrupBombTag extends BattlerTag {
-  constructor(sourceId: number) {
-    super(BattlerTagType.SYRUP_BOMB, BattlerTagLapseType.TURN_END, 3, MoveId.SYRUP_BOMB, sourceId);
-  }
-
-  /**
-   * Adds the Syrup Bomb battler tag to the target Pokemon.
-   * @param pokemon - The target {@linkcode Pokemon}
-   */
-  override onAdd(pokemon: Pokemon) {
-    super.onAdd(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:syrupBombOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Applies the single-stage speed down to the target Pokemon and decrements the tag's turn count
-   * @param pokemon - The target {@linkcode Pokemon}
-   * @param _lapseType - N/A
-   * @returns `true` if the `turnCount` is still greater than `0`; `false` if the `turnCount` is `0` or the target or source Pokemon has been removed from the field
-   */
-  override lapse(pokemon: Pokemon, _lapseType: BattlerTagLapseType): boolean {
-    if (this.sourceId && !globalScene.getPokemonById(this.sourceId)?.isActive(true)) {
-      return false;
-    }
-    // Custom message in lieu of an animation in mainline
-    globalScene.queueMessage(
-      i18next.t("battlerTags:syrupBombLapse", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-    globalScene.unshiftPhase(
-      new StatStageChangePhase(pokemon.getBattlerIndex(), null, [Stat.SPD], -1, { bypassReflect: true }),
-    );
-    return --this.turnCount > 0;
-  }
-}
-
-/**
- * Telekinesis raises the target into the air for three turns and causes all moves used against the target (aside from OHKO moves) to hit the target unless the target is in a semi-invulnerable state from Fly/Dig.
- * The first effect is provided by {@linkcode FloatingTag}, the accuracy-bypass effect is provided by TelekinesisTag
- * The effects of Telekinesis can be baton passed to a teammate. Unlike the mainline games, Telekinesis can be baton-passed to Mega Gengar.
- * @see {@link https://bulbapedia.bulbagarden.net/wiki/Telekinesis_(move) | MoveId.TELEKINESIS}
- */
-export class TelekinesisTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId) {
-    super(
-      BattlerTagType.TELEKINESIS,
-      [BattlerTagLapseType.PRE_MOVE, BattlerTagLapseType.AFTER_MOVE],
-      3,
-      sourceMoveId,
-      undefined,
-      true,
-    );
-  }
-
-  override onAdd(pokemon: Pokemon) {
-    globalScene.queueMessage(
-      i18next.t("battlerTags:telekinesisOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-}
-
-/**
- * Tag that swaps the user's base ATK stat with its base DEF stat.
- * @extends BattlerTag
- */
-export class PowerTrickTag extends BattlerTag {
-  constructor(sourceMoveId: MoveId, sourceId: number) {
-    super(BattlerTagType.POWER_TRICK, BattlerTagLapseType.CUSTOM, 0, sourceMoveId, sourceId, true);
-  }
-
-  override onAdd(pokemon: Pokemon): void {
-    this.swapStat(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:powerTrickActive", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  override onRemove(pokemon: Pokemon): void {
-    this.swapStat(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:powerTrickActive", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Removes the Power Trick tag and reverts any stat changes if the tag is already applied.
-   * @param pokemon The {@linkcode Pokemon} that already has the Power Trick tag.
-   */
-  override onOverlap(pokemon: Pokemon): void {
-    pokemon.removeTag(this.tagType);
-  }
-
-  /**
-   * Swaps the user's base ATK stat with its base DEF stat.
-   * @param pokemon The {@linkcode Pokemon} whose stats will be swapped.
-   */
-  swapStat(pokemon: Pokemon): void {
-    const temp = pokemon.getStat(Stat.ATK, false);
-    pokemon.setStat(Stat.ATK, pokemon.getStat(Stat.DEF, false), false);
-    pokemon.setStat(Stat.DEF, temp, false);
-  }
-}
-
-/**
- * Tag associated with the move Grudge.
- * If this tag is active when the bearer faints from an opponent's move, the tag reduces that move's PP to 0.
- * Otherwise, it lapses when the bearer makes another move.
- */
-export class GrudgeTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.GRUDGE, [BattlerTagLapseType.CUSTOM, BattlerTagLapseType.PRE_MOVE], 1, MoveId.GRUDGE);
-  }
-
-  override onAdd(pokemon: Pokemon) {
-    super.onAdd(pokemon);
-    globalScene.queueMessage(
-      i18next.t("battlerTags:grudgeOnAdd", { pokemonNameWithAffix: getPokemonNameWithAffix(pokemon) }),
-    );
-  }
-
-  /**
-   * Activates Grudge's special effect on the attacking Pokemon and lapses the tag.
-   * @param pokemon
-   * @param lapseType
-   * @param sourcePokemon {@linkcode Pokemon} the source of the move that fainted the tag's bearer
-   * @returns `false` if Grudge activates its effect or lapses
-   */
-  override lapse(pokemon: Pokemon, lapseType: BattlerTagLapseType, sourcePokemon?: Pokemon): boolean {
-    if (lapseType === BattlerTagLapseType.CUSTOM && sourcePokemon) {
-      if (sourcePokemon.isActive() && pokemon.isOpponent(sourcePokemon)) {
-        const lastMove = pokemon.turnData.attacksReceived[0];
-        const lastMoveData = sourcePokemon.getMoveset().find((m) => m.moveId === lastMove.moveId);
-        if (lastMoveData && lastMove.moveId !== MoveId.STRUGGLE) {
-          lastMoveData.ppUsed = lastMoveData.getMovePp();
-          globalScene.queueMessage(
-            i18next.t("battlerTags:grudgeLapse", {
-              pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-              moveName: lastMoveData.getName(),
-            }),
-          );
-        }
-      }
-      return false;
-    } else {
-      return super.lapse(pokemon, lapseType);
-    }
-  }
-}
-
-/**
- * Tag used to heal the user of Psycho Shift of its status effect if Psycho Shift succeeds in transferring its status effect to the target Pokemon
- */
-export class PsychoShiftTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.PSYCHO_SHIFT, BattlerTagLapseType.AFTER_MOVE, 1, MoveId.PSYCHO_SHIFT);
-  }
-
-  /**
-   * Heals Psycho Shift's user of its status effect after it uses a move
-   * @returns `false` to expire the tag immediately
-   */
-  override lapse(pokemon: Pokemon, _lapseType: BattlerTagLapseType): boolean {
-    if (pokemon.hasNonVolatileStatusEffect() && pokemon.isActive(true)) {
-      globalScene.queueMessage(getStatusEffectHealText(pokemon.getStatusEffect(), getPokemonNameWithAffix(pokemon)));
-      pokemon.resetStatus();
-      pokemon.updateInfo();
-    }
-    return false;
-  }
-}
-
-/**
- * Tag to allow the affected Pokemon's move to go first in its priority bracket.
- * Used for {@link https://bulbapedia.bulbagarden.net/wiki/Quick_Draw_(Ability) Quick Draw}
- * and {@link https://bulbapedia.bulbagarden.net/wiki/Quick_Claw Quick Claw}.
- */
-export class BypassSpeedTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.BYPASS_SPEED, BattlerTagLapseType.TURN_END, 1);
-  }
-
-  override canAdd(pokemon: Pokemon): boolean {
-    const cancelled = new BooleanHolder(false);
-    applyAbAttrs(AbAttrFlag.PREVENT_BYPASS_SPEED_CHANCE, pokemon, false, cancelled);
-    return !cancelled.value;
-  }
-}
-
-/**
- * Tag to force the affected Pokemon to move last in the turn.
- * Used for {@link https://bulbapedia.bulbagarden.net/wiki/Quash_(move) | Quash}.
- */
-export class QuashedTag extends BattlerTag {
-  constructor() {
-    super(BattlerTagType.QUASHED, BattlerTagLapseType.TURN_END, 1);
-  }
-
-  override onAdd(pokemon: Pokemon) {
-    // "{pokemonNameWithAffix}'s move was postponed!"
-    globalScene.queueMessage(
-      i18next.t("battlerTags:quashOnAdd", {
-        pokemonNameWithAffix: getPokemonNameWithAffix(pokemon),
-      }),
-    );
-  }
-}
-
-/**
- * Retrieves a {@linkcode BattlerTag} based on the provided tag type, turn count, source move, and source ID.
- * @param sourceId - The ID of the pokemon adding the tag
- * @returns The corresponding {@linkcode BattlerTag} object.
- */
-export function getBattlerTag(
-  tagType: BattlerTagType,
-  turnCount: number,
-  sourceMoveId: MoveId,
-  sourceId: number,
-): BattlerTag {
-  switch (tagType) {
-    case BattlerTagType.RECHARGING:
-      return new RechargingTag(sourceMoveId);
-    case BattlerTagType.BEAK_BLAST_CHARGING:
-      return new BeakBlastChargingTag();
-    case BattlerTagType.RAGE:
-      return new RageTag();
-    case BattlerTagType.SHELL_TRAP:
-      return new ShellTrapTag();
-    case BattlerTagType.FLINCHED:
-      return new FlinchedTag(sourceMoveId);
-    case BattlerTagType.INTERRUPTED:
-      return new InterruptedTag(sourceMoveId);
-    case BattlerTagType.CONFUSED:
-      return new ConfusedTag(turnCount, sourceMoveId);
-    case BattlerTagType.INFATUATED:
-      return new InfatuatedTag(sourceMoveId, sourceId);
-    case BattlerTagType.SEEDED:
-      return new SeedTag(sourceId);
-    case BattlerTagType.POWDER:
-      return new PowderTag();
-    case BattlerTagType.NIGHTMARE:
-      return new NightmareTag();
-    case BattlerTagType.FRENZY:
-      return new FrenzyTag(turnCount, sourceMoveId);
-    case BattlerTagType.ROLLING:
-      return new RollingTag(sourceMoveId);
-    case BattlerTagType.UPROAR:
-      return new UproarTag();
-    case BattlerTagType.CHARGING:
-      return new BattlerTag(tagType, BattlerTagLapseType.CUSTOM, 1, sourceMoveId, sourceId);
-    case BattlerTagType.ENCORE:
-      return new EncoreTag(sourceId);
-    case BattlerTagType.HELPING_HAND:
-      return new HelpingHandTag(sourceId);
-    case BattlerTagType.INGRAIN:
-      return new IngrainTag(sourceId);
-    case BattlerTagType.AQUA_RING:
-      return new AquaRingTag();
-    case BattlerTagType.DROWSY:
-      return new DrowsyTag();
-    case BattlerTagType.TRAPPED:
-      return new TrappedTag(tagType, BattlerTagLapseType.CUSTOM, turnCount, sourceMoveId, sourceId);
-    case BattlerTagType.NO_RETREAT:
-      return new NoRetreatTag(sourceId);
-    case BattlerTagType.BIND:
-      return new BindTag(turnCount, sourceId);
-    case BattlerTagType.WRAP:
-      return new WrapTag(turnCount, sourceId);
-    case BattlerTagType.FIRE_SPIN:
-      return new FireSpinTag(turnCount, sourceId);
-    case BattlerTagType.G_MAX_FIRE_SPIN:
-      return new GMaxFireSpinTag(turnCount, sourceId);
-    case BattlerTagType.WHIRLPOOL:
-      return new WhirlpoolTag(turnCount, sourceId);
-    case BattlerTagType.CLAMP:
-      return new ClampTag(turnCount, sourceId);
-    case BattlerTagType.SAND_TOMB:
-      return new SandTombTag(turnCount, sourceId);
-    case BattlerTagType.G_MAX_SAND_TOMB:
-      return new GMaxSandTombTag(turnCount, sourceId);
-    case BattlerTagType.MAGMA_STORM:
-      return new MagmaStormTag(turnCount, sourceId);
-    case BattlerTagType.SNAP_TRAP:
-      return new SnapTrapTag(turnCount, sourceId);
-    case BattlerTagType.THUNDER_CAGE:
-      return new ThunderCageTag(turnCount, sourceId);
-    case BattlerTagType.INFESTATION:
-      return new InfestationTag(turnCount, sourceId);
-    case BattlerTagType.PROTECTED:
-      return new ProtectedTag(sourceMoveId);
-    case BattlerTagType.SPIKY_SHIELD:
-      return new ContactDamageProtectedTag(sourceMoveId, 8);
-    case BattlerTagType.KINGS_SHIELD:
-      return new ContactStatStageChangeProtectedTag(sourceMoveId, tagType, Stat.ATK, -1);
-    case BattlerTagType.OBSTRUCT:
-      return new ContactStatStageChangeProtectedTag(sourceMoveId, tagType, Stat.DEF, -2);
-    case BattlerTagType.SILK_TRAP:
-      return new ContactStatStageChangeProtectedTag(sourceMoveId, tagType, Stat.SPD, -1);
-    case BattlerTagType.BANEFUL_BUNKER:
-      return new ContactPoisonProtectedTag(sourceMoveId);
-    case BattlerTagType.BURNING_BULWARK:
-      return new ContactBurnProtectedTag(sourceMoveId);
-    case BattlerTagType.ENDURING:
-      return new EnduringTag(tagType, BattlerTagLapseType.TURN_END, sourceMoveId);
-    case BattlerTagType.MAGIC_COAT:
-      return new MagicCoatTag();
-    case BattlerTagType.STURDY:
-      return new SturdyTag(sourceMoveId);
-    case BattlerTagType.PERISH_SONG:
-      return new PerishSongTag(turnCount);
-    case BattlerTagType.CENTER_OF_ATTENTION:
-      return new CenterOfAttentionTag(sourceMoveId);
-    case BattlerTagType.TRUANT:
-      return new TruantTag();
-    case BattlerTagType.SLOW_START:
-      return new SlowStartTag();
-    case BattlerTagType.PROTOSYNTHESIS:
-      return new WeatherHighestStatBoostTag(
-        tagType,
-        Abilities.PROTOSYNTHESIS,
-        WeatherType.SUNNY,
-        WeatherType.HARSH_SUN,
-      );
-    case BattlerTagType.QUARK_DRIVE:
-      return new TerrainHighestStatBoostTag(tagType, Abilities.QUARK_DRIVE, TerrainType.ELECTRIC);
-    case BattlerTagType.FLYING:
-    case BattlerTagType.UNDERGROUND:
-    case BattlerTagType.UNDERWATER:
-    case BattlerTagType.HIDDEN:
-      return new SemiInvulnerableTag(tagType, turnCount, sourceMoveId);
-    case BattlerTagType.SKY_DROP:
-      return new SkyDropTag(sourceId);
-    case BattlerTagType.FIRE_BOOST:
-      return new TypeBoostTag(tagType, sourceMoveId, ElementalType.FIRE, 1.5, false);
-    case BattlerTagType.CRIT_BOOST:
-      return new CritBoostTag(tagType, sourceMoveId);
-    case BattlerTagType.DRAGON_CHEER:
-      return new DragonCheerTag();
-    case BattlerTagType.CRIT_BOOST_STACKABLE:
-      return new CritBoostStackableTag();
-    case BattlerTagType.ALWAYS_CRIT:
-    case BattlerTagType.IGNORE_ACCURACY:
-      return new BattlerTag(tagType, BattlerTagLapseType.TURN_END, 2, sourceMoveId);
-    case BattlerTagType.ALWAYS_GET_HIT:
-    case BattlerTagType.RECEIVE_DOUBLE_DAMAGE:
-      return new BattlerTag(tagType, BattlerTagLapseType.PRE_MOVE, 1, sourceMoveId);
-    case BattlerTagType.BYPASS_SLEEP:
-      return new BattlerTag(tagType, BattlerTagLapseType.TURN_END, turnCount, sourceMoveId);
-    case BattlerTagType.IGNORE_FLYING:
-      return new GroundedTag(tagType, BattlerTagLapseType.CUSTOM, sourceMoveId);
-    case BattlerTagType.ROOSTED:
-      return new RoostedTag();
-    case BattlerTagType.BURNED_UP:
-      return new RemovedTypeTag(tagType, BattlerTagLapseType.CUSTOM, sourceMoveId);
-    case BattlerTagType.DOUBLE_SHOCKED:
-      return new RemovedTypeTag(tagType, BattlerTagLapseType.CUSTOM, sourceMoveId);
-    case BattlerTagType.SALT_CURED:
-      return new SaltCuredTag(sourceId);
-    case BattlerTagType.CURSED:
-      return new CursedTag(sourceId);
-    case BattlerTagType.CHARGED:
-      return new TypeBoostTag(tagType, sourceMoveId, ElementalType.ELECTRIC, 2, true);
-    case BattlerTagType.FLOATING:
-      return new FloatingTag(tagType, sourceMoveId, turnCount);
-    case BattlerTagType.MINIMIZED:
-      return new MinimizeTag();
-    case BattlerTagType.DESTINY_BOND:
-      return new DestinyBondTag(sourceMoveId, sourceId);
-    case BattlerTagType.ICE_FACE:
-      return new IceFaceBlockDamageTag(tagType);
-    case BattlerTagType.DISGUISE:
-      return new FormBlockDamageTag(tagType);
-    case BattlerTagType.COMMANDED:
-      return new CommandedTag(sourceId);
-    case BattlerTagType.STOCKPILING:
-      return new StockpilingTag(sourceMoveId);
-    case BattlerTagType.OCTOLOCK:
-      return new OctolockTag(sourceId);
-    case BattlerTagType.DISABLED:
-      return new DisabledTag(sourceId);
-    case BattlerTagType.IGNORE_GHOST:
-      return new ExposedTag(tagType, sourceMoveId, ElementalType.GHOST, [ElementalType.NORMAL, ElementalType.FIGHTING]);
-    case BattlerTagType.IGNORE_DARK:
-      return new ExposedTag(tagType, sourceMoveId, ElementalType.DARK, [ElementalType.PSYCHIC]);
-    case BattlerTagType.GULP_MISSILE_ARROKUDA:
-    case BattlerTagType.GULP_MISSILE_PIKACHU:
-      return new GulpMissileTag(tagType, sourceMoveId);
-    case BattlerTagType.TAR_SHOT:
-      return new TarShotTag();
-    case BattlerTagType.ELECTRIFIED:
-      return new ElectrifiedTag();
-    case BattlerTagType.THROAT_CHOPPED:
-      return new ThroatChoppedTag();
-    case BattlerTagType.GORILLA_TACTICS:
-      return new GorillaTacticsTag();
-    case BattlerTagType.UNBURDEN:
-      return new UnburdenTag();
-    case BattlerTagType.SUBSTITUTE:
-      return new SubstituteTag(sourceMoveId, sourceId);
-    case BattlerTagType.AUTOTOMIZED:
-      return new AutotomizedTag();
-    case BattlerTagType.MYSTERY_ENCOUNTER_POST_SUMMON:
-      return new MysteryEncounterPostSummonTag();
-    case BattlerTagType.HEAL_BLOCK:
-      return new HealBlockTag(turnCount, sourceMoveId);
-    case BattlerTagType.TORMENT:
-      return new TormentTag(sourceId);
-    case BattlerTagType.TAUNT:
-      return new TauntTag();
-    case BattlerTagType.IMPRISONING:
-      return new ImprisoningTag();
-    case BattlerTagType.SYRUP_BOMB:
-      return new SyrupBombTag(sourceId);
-    case BattlerTagType.TELEKINESIS:
-      return new TelekinesisTag(sourceMoveId);
-    case BattlerTagType.POWER_TRICK:
-      return new PowerTrickTag(sourceMoveId, sourceId);
-    case BattlerTagType.GRUDGE:
-      return new GrudgeTag();
-    case BattlerTagType.PSYCHO_SHIFT:
-      return new PsychoShiftTag();
-    case BattlerTagType.BYPASS_SPEED:
-      return new BypassSpeedTag();
-    case BattlerTagType.QUASHED:
-      return new QuashedTag();
-    case BattlerTagType.NONE:
-    default:
-      return new BattlerTag(tagType, BattlerTagLapseType.CUSTOM, turnCount, sourceMoveId, sourceId);
-  }
-}
-
-/**
- * When given a battler tag or json representing one, creates an actual BattlerTag object with the same data.
- * @param source A battler tag
- * @return The valid battler tag
- */
-export function loadBattlerTag(source: BattlerTag | any): BattlerTag {
-  const tag = getBattlerTag(source.tagType, source.turnCount, source.sourceMoveId, source.sourceId);
-  tag.loadTag(source);
-  return tag;
-}
-
-/**
- * Helper function to verify that the current phase is a MoveEffectPhase and provide quick access to commonly used fields
- *
- * @param pokemon {@linkcode Pokemon} The Pokémon used to access the current phase
- * @returns null if current phase is not MoveEffectPhase, otherwise Object containing the {@linkcode MoveEffectPhase}, and its
- * corresponding {@linkcode Move} and user {@linkcode Pokemon}
- */
-function getMoveEffectPhaseData(_pokemon: Pokemon): { phase: MoveEffectPhase; attacker: Pokemon; move: Move } | null {
-  const phase = globalScene.getCurrentPhase();
-  if (phase?.is<MoveEffectPhase>(PhaseId.MOVE_EFFECT)) {
-    return {
-      phase: phase,
-      attacker: phase.getPokemon(),
-      move: phase.move.getMove(),
-    };
-  }
-  return null;
-}
