+++ src/phases/move-phase.ts
@@ -1,9 +1,9 @@
 import type { RedirectMoveAbAttr } from "#app/data/abilities/ab-attrs/redirect-move-ab-attr";
 import type { ReflectMovesAbAttr } from "#app/data/abilities/ab-attrs/reflect-moves-ab-attr";
 import { applyAbAttrs } from "#app/data/abilities/apply-ab-attrs";
-import type { MagicCoatTag } from "#app/data/battler-tags";
 import type { CenterOfAttentionTag } from "#app/data/battler-tags/center-of-attention-tag";
 import type { ImprisoningTag } from "#app/data/battler-tags/imprisoning-tag";
+import type { MagicCoatTag } from "#app/data/battler-tags/magic-coat-tag";
 import { applyBattlerTags } from "#app/data/battler-tags/utils/apply-battler-tags";
 import { allMoves } from "#app/data/data-lists";
 import { getMoveTargets, SelfStatusMove } from "#app/data/moves/move";
