+++ test/moves/magic-coat.test.ts
@@ -1,4 +1,4 @@
-import type { EncoreTag } from "#app/data/battler-tags";
+import type { EncoreTag } from "#app/data/battler-tags/encore-tag";
 import { allAbilities, allMoves } from "#app/data/data-lists";
 import { Abilities } from "#enums/abilities";
 import { ArenaTagSide } from "#enums/arena-tag-side";
